% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_bayesian_ocs.R
\name{average_power}
\alias{average_power}
\title{Calculate the average power}
\usage{
average_power(prepost_proba_TP, prior_proba_no_benefit)
}
\arguments{
\item{prepost_proba_TP}{The preposterior probability of true positive.}

\item{prior_proba_no_benefit}{The prior probability of no benefit.}
}
\value{
The average power.
}
\description{
This function calculates the average power based on the preposterior probability of true positive
and the prior probability of no benefit.
}
\examples{
NA
}
