% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_bayesian_ocs.R
\name{average_tie}
\alias{average_tie}
\title{Calculate the average type 1 error}
\usage{
average_tie(prepost_proba_FP, prior_proba_no_benefit)
}
\arguments{
\item{prepost_proba_FP}{The preposterior probability of false positive.}

\item{prior_proba_no_benefit}{The prior probability of no benefit.}
}
\value{
The average type 1 error.
}
\description{
This function calculates the average type 1 error based on the preposterior probability of false positive
and the prior probability of no benefit.
}
\examples{
NA
}
