% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian_operating_characteristics.R
\name{bayesian_metric_vs_parameters}
\alias{bayesian_metric_vs_parameters}
\title{Function to plot metric vs parameters}
\usage{
bayesian_metric_vs_parameters(
  results_metrics_df,
  metric,
  case_study = "belimumab",
  method = "RMP",
  target_sample_size_per_arm = 93,
  theta_0 = NULL,
  add_baselines = FALSE,
  source_denominator_change_factor = 1,
  target_to_source_std_ratio = 1
)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metric}{The metric to plot.}

\item{case_study}{The case study to plot the metric for.}

\item{method}{The method to plot the metric for.}

\item{target_sample_size_per_arm}{The target sample size per arm to plot the metric for.}

\item{theta_0}{Boundary of the null hypothesis space.}

\item{add_baselines}{A logical value indicating whether to add baselines to the plot.}
}
\value{
None
}
\description{
This function plots a metric against the parameters for a given method, case study, and sample size.
}
