% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian_operating_characteristics.R
\name{bayesian_ocs_plots}
\alias{bayesian_ocs_plots}
\title{Plot methods operating characteristics}
\usage{
bayesian_ocs_plots(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of different methods.
}
\examples{
NA
}
