% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_bayesian_ocs.R
\name{bayesian_ocs_scenario_simulation}
\alias{bayesian_ocs_scenario_simulation}
\title{Simulate a Scenario}
\usage{
bayesian_ocs_scenario_simulation(
  scenario,
  simulation_config,
  scenarios_config,
  bayes_filename,
  config_dir,
  case_studies_config_dir
)
}
\arguments{
\item{scenario}{The scenario to simulate.}

\item{simulation_config}{Simulation configuration}

\item{scenarios_config}{Scenario-generation configuration}

\item{bayes_filename}{Results were results are stored for Bayesian OCs}

\item{config_dir}{Configuration files directory}

\item{case_studies_config_dir}{Case-study configuration directory}
}
\value{
The results of the simulation.
}
\description{
This function simulates a scenario based on the provided inputs and returns the results of the simulation.
}
