% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{combine_parameters}
\alias{combine_parameters}
\title{Combine parameters for a given method.}
\usage{
combine_parameters(method)
}
\arguments{
\item{method}{The method name.}
}
\value{
A list of combinations of parameters for the method.
}
\description{
Combine parameters for a given method.
}
