% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorised_commensurate_prior.R
\name{commensurate_tau_quadrature}
\alias{commensurate_tau_quadrature}
\title{Quadrature rule for the commensurability parameter}
\usage{
commensurate_tau_quadrature(model, n_nodes)
}
\arguments{
\item{model}{A \link{GaussianCommensuratePowerPrior} object.}

\item{n_nodes}{Number of nodes.}
}
\value{
A list with \code{tau}, \code{inverse_tau}, \code{log_tau} and \code{weights}.
}
\description{
Quadrature rule for the commensurability parameter
}
\keyword{internal}
