% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{compute_control_drift_range}
\alias{compute_control_drift_range}
\title{Compute the control drift range for a given source treatment effect.}
\usage{
compute_control_drift_range(source_treatment_effect)
}
\arguments{
\item{source_treatment_effect}{The source treatment effect.}
}
\value{
A vector representing the control drift range.
}
\description{
Compute the control drift range for a given source treatment effect.
}
