% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{compute_drift_range}
\alias{compute_drift_range}
\title{Compute the drift range for a given simulation and case study configuration.}
\usage{
compute_drift_range(scenarios_config, case_study_config)
}
\arguments{
\item{scenarios_config}{Configuration for the simulation.}

\item{case_study_config}{Configuration for the case study.}
}
\value{
A vector representing the drift range.
}
\description{
Compute the drift range for a given simulation and case study configuration.
}
