% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_operating_characteristics.R
\name{compute_freq_power_pooling}
\alias{compute_freq_power_pooling}
\title{Compute the frequentist power}
\usage{
compute_freq_power_pooling(
  alpha,
  target_data,
  source_data,
  frequentist_test,
  theta_0,
  null_space,
  simulation_config,
  case_study = NULL,
  n_replicates = 1000
)
}
\arguments{
\item{alpha}{The significance level.}

\item{target_data}{Target data object}

\item{source_data}{Source study data}

\item{frequentist_test}{Type of frequentist test to apply, either z-test or t-test}

\item{theta_0}{Boundary of the null hypothesis space}

\item{null_space}{Side of the null space, either left or right.}

\item{simulation_config}{Simulation configuration.}

\item{case_study}{Optional case-study name.}

\item{n_replicates}{Number of Monte Carlo replicates for non-analytical power calculations.}
}
\value{
The power of the test.
}
\description{
This function computes the power of a test for a pooled analysis at a given significance level
}
