% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{compute_log_odds_ratio_from_counts}
\alias{compute_log_odds_ratio_from_counts}
\title{Compute the log odds ratio from count data}
\usage{
compute_log_odds_ratio_from_counts(
  n_control_responders,
  n_treatment_responders,
  n_control_nonresponders,
  n_treatment_nonresponders,
  continuity_correction
)
}
\arguments{
\item{n_control_responders}{Number of responders in the control arm}

\item{n_treatment_responders}{Number of responders in the treatment arm}

\item{n_control_nonresponders}{Number of nonresponders in the control arm}

\item{n_treatment_nonresponders}{Number of nonresponders in the treatment arm}

\item{continuity_correction}{Whether to apply continuity correction or not}
}
\value{
The log odds ratio
}
\description{
Compute the log odds ratio from count data
}
\keyword{internal}
