% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{compute_log_odds_ratio_from_rates}
\alias{compute_log_odds_ratio_from_rates}
\title{Compute the log odds ratio based on two rates}
\usage{
compute_log_odds_ratio_from_rates(rate1, rate2)
}
\arguments{
\item{rate1}{The rate in the first group}

\item{rate2}{The rate in the second group}
}
\value{
The log odds ratio
}
\description{
This function computes the log odds ratio based on two rates.
}
\keyword{internal}
