% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{compute_ORs}
\alias{compute_ORs}
\title{Compute odds ratios based on the number of participants and responders in each arm}
\usage{
compute_ORs(
  n_control,
  n_treatment,
  n_control_responders,
  n_treatment_responders
)
}
\arguments{
\item{n_control}{Number of participants in the control arm}

\item{n_treatment}{Number of participants in the treatment arm}

\item{n_control_responders}{Number of responders in the control arm}

\item{n_treatment_responders}{Number of responders in the treatment arm}
}
\value{
A list containing the log odds ratio, treatment rate, control rate, and standard error of the log odds ratio
}
\description{
This function computes odds ratios based on the number of participants and responders in the control and treatment arms.
}
\keyword{internal}
