% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{compute_source_denominator_range}
\alias{compute_source_denominator_range}
\title{Compute the source denominator range for a given source data, simulation configuration, and case study configuration.}
\usage{
compute_source_denominator_range(
  source_data,
  scenarios_config,
  case_study_config
)
}
\arguments{
\item{source_data}{The source data.}

\item{scenarios_config}{Configuration for the simulation.}

\item{case_study_config}{Configuration for the case study.}
}
\value{
A vector representing the source denominator range.
}
\description{
Compute the source denominator range for a given source data, simulation configuration, and case study configuration.
}
