% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_simulation_logs}
\alias{concatenate_simulation_logs}
\title{Concatenate all simulation logs to a global dataframe}
\usage{
concatenate_simulation_logs(results_dir)
}
\arguments{
\item{results_dir}{A character string specifying the path of the results folder.}
}
\value{
No return value, called for side effects.
}
\description{
This function reads all CSV files with names starting with "logs" from both the "bayesian" and
"frequentist" subdirectories, concatenates them, adds an "OCs" column, and saves the result to a new CSV file.
}
