% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concatenate_simulation_results}
\alias{concatenate_simulation_results}
\title{Concatenate all simulation results to a global dataframe}
\usage{
concatenate_simulation_results(results_dir, ocs_filename)
}
\arguments{
\item{results_dir}{A character string specifying the path of the results folder.}

\item{ocs_filename}{A character string specifying the filename of the results dataframe.}
}
\value{
No return value, called for side effects.
}
\description{
This function reads all CSV files with names starting with "results" from the specified
directory, concatenates them, and saves the result to a new CSV file.
}
\keyword{internal}
