% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validation.R
\name{config_type_descriptions}
\alias{config_type_descriptions}
\title{Describe what a configuration type requires, for error messages}
\usage{
config_type_descriptions
}
\description{
Describe what a configuration type requires, for error messages
}
\keyword{internal}
