% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{convert_params_to_str}
\alias{convert_params_to_str}
\title{Function to convert parameters in dataframe to string}
\usage{
convert_params_to_str(method, parameters)
}
\arguments{
\item{method}{The method name.}

\item{parameters}{The parameters dataframe.}
}
\value{
The string representation of the parameters.
}
\description{
Function to convert parameters in dataframe to string
}
\keyword{internal}
