% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_empirical_bayes_hyperparameters_vs_scenario.R
\name{empirical_bayes_parameters_plots}
\alias{empirical_bayes_parameters_plots}
\title{Plot the posterior parameters}
\usage{
empirical_bayes_parameters_plots(results_metrics_df)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}
}
\value{
None
}
\description{
Plot the hyperparameters updated using empirical Bayes as a function of drift, for each scenario.
}
