% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods_comparison_forestplot.R
\name{forest_plot_methods_comparison}
\alias{forest_plot_methods_comparison}
\title{Plot methods comparison}
\usage{
forest_plot_methods_comparison(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The results metrics dataframe.}

\item{metrics}{Metrics for which to create forest plots.}
}
\value{
None
}
\description{
This function plots the comparison of methods based on the provided results metrics dataframe.
}
