% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sweet_spots_comparison.R
\name{forest_plot_sweet_spot}
\alias{forest_plot_sweet_spot}
\title{Generate a forest plot}
\usage{
forest_plot_sweet_spot(results_freq_df, x_metric)
}
\arguments{
\item{results_freq_df}{The results dataframe.}

\item{x_metric}{Metric on the x-axis}

\item{selected_case_study}{The selected case study.}

\item{selected_target_sample_size_per_arm}{The selected target sample size per arm.}
}
\value{
None
}
\description{
This function generates a forest plot based on the provided results dataframe, selected case study, and selected target sample size per arm.
}
