% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sweet_spots_comparison.R
\name{forest_plot_sweet_spots_comparison}
\alias{forest_plot_sweet_spots_comparison}
\title{Plot methods comparison}
\usage{
forest_plot_sweet_spots_comparison(sweet_spots_df, metrics)
}
\arguments{
\item{sweet_spots_df}{The results metrics dataframe.}

\item{metrics}{Metrics for which to create forest plots.}
}
\value{
None
}
\description{
This function plots the comparison of methods based on the provided results metrics dataframe.
}
