% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{format_uncertainty}
\alias{format_uncertainty}
\title{Function to format uncertainty}
\usage{
format_uncertainty(yerr_input, y, metric)
}
\arguments{
\item{yerr_input}{The input uncertainty values.}

\item{y}{The central values.}

\item{metric}{The metric information.}
}
\value{
The formatted uncertainty.
}
\description{
Function to format uncertainty
}
\keyword{internal}
