% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_operating_characteristics.R
\name{frequentist_power_at_equivalent_tie}
\alias{frequentist_power_at_equivalent_tie}
\title{Compute the frequentist power at equivalent tie}
\usage{
frequentist_power_at_equivalent_tie(
  results,
  analysis_config,
  simulation_config,
  parallelization = FALSE
)
}
\arguments{
\item{results}{The results data frame.}

\item{analysis_config}{The analysis configuration.}
}
\value{
The final results data frame with power and frequentist test columns added.
}
\description{
This function computes the frequentist power at equivalent tie for a given set of results and analysis configuration.
}
