% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_prior_ess_estimation.R
\name{gaussian_mix_moment_ess}
\alias{gaussian_mix_moment_ess}
\title{Calculate the moment-based Effective Sample Size (ESS) for a Gaussian mixture.}
\usage{
gaussian_mix_moment_ess(mix)
}
\arguments{
\item{mix}{A mixture object.}
}
\value{
The moment-based ESS for the Gaussian mixture.
}
\description{
This function calculates the moment-based ESS for a Gaussian mixture distribution.
It takes a mixture object and an optional prior reference scale as input and returns the ESS.
}
