% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_elastic_prior.R
\name{GaussianElasticPriorStep}
\alias{GaussianElasticPriorStep}
\title{GaussianElasticPriorStep class}
\value{
An R6 class object representing a GaussianElasticPriorStep model
}
\description{
This class represents an Elastic Prior model.
It inherits from the Model class.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:ConjugateGaussian]{RBExT::ConjugateGaussian}} -> \code{GaussianElasticPriorStep}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{congruence_threshold}}{Congruence threshold}

\item{\code{target_standard_error}}{Standard error on the treatment effect in the target study}

\item{\code{target_treatment_effect_estimate}}{Estimate of the treatment effect in the target study}

\item{\code{target_sample_size_per_arm}}{Sample size per arm in the target study}

\item{\code{source_sample_size_per_arm}}{Sample size per arm in the source study}

\item{\code{source_data_variance}}{Estimated sampling variance in the source study}

\item{\code{method}}{Name of the method}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianElasticPriorStep-new}{\code{GaussianElasticPriorStep$new()}}
\item \href{#method-GaussianElasticPriorStep-empirical_bayes_update}{\code{GaussianElasticPriorStep$empirical_bayes_update()}}
\item \href{#method-GaussianElasticPriorStep-method_calibration}{\code{GaussianElasticPriorStep$method_calibration()}}
\item \href{#method-GaussianElasticPriorStep-generate_congruence_measure_distribution}{\code{GaussianElasticPriorStep$generate_congruence_measure_distribution()}}
\item \href{#method-GaussianElasticPriorStep-setup_prior}{\code{GaussianElasticPriorStep$setup_prior()}}
\item \href{#method-GaussianElasticPriorStep-clone}{\code{GaussianElasticPriorStep$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="inference"><a href='../../RBExT/html/Model.html#method-Model-inference'><code>RBExT::Model$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="credible_interval"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-credible_interval'><code>RBExT::ConjugateGaussian$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_cdf'><code>RBExT::ConjugateGaussian$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_mean"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_mean'><code>RBExT::ConjugateGaussian$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_median"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_median'><code>RBExT::ConjugateGaussian$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_moments"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_moments'><code>RBExT::ConjugateGaussian$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_pdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_pdf'><code>RBExT::ConjugateGaussian$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_to_RBesT'><code>RBExT::ConjugateGaussian$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_variance"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_variance'><code>RBExT::ConjugateGaussian$posterior_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_cdf'><code>RBExT::ConjugateGaussian$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_pdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_pdf'><code>RBExT::ConjugateGaussian$prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_to_RBesT'><code>RBExT::ConjugateGaussian$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_posterior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_posterior'><code>RBExT::ConjugateGaussian$sample_posterior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_prior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_prior'><code>RBExT::ConjugateGaussian$sample_prior()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-new}{}}}
\subsection{Method \code{new()}}{
Initializes the GaussianElasticPriorStep object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$new(prior)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior information for the analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-empirical_bayes_update"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-empirical_bayes_update}{}}}
\subsection{Method \code{empirical_bayes_update()}}{
Updates the vague prior variance based on empirical Bayes approach.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$empirical_bayes_update(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-method_calibration"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-method_calibration}{}}}
\subsection{Method \code{method_calibration()}}{
Calibrates the method based on target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$method_calibration(
  target_sample_size_per_arm,
  target_standard_error,
  target_treatment_effect_estimate
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_sample_size_per_arm}}{Sample size per arm in the target study}

\item{\code{target_standard_error}}{Standard error of the treatment effect in the target study}

\item{\code{target_treatment_effect_estimate}}{Estimate of the treatment effect in the target study}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-generate_congruence_measure_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-generate_congruence_measure_distribution}{}}}
\subsection{Method \code{generate_congruence_measure_distribution()}}{
Generate distribution of congruence measure T in the homogeneous case
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$generate_congruence_measure_distribution(
  target_sample_size_per_arm,
  target_treatment_effect_estimate,
  target_data_variance
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_sample_size_per_arm}}{Sample size per arm in the target study}

\item{\code{target_treatment_effect_estimate}}{Treatment effect estimate in the target study}

\item{\code{target_data_variance}}{Estimated sampling variance of the target data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of congruence measures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-setup_prior"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-setup_prior}{}}}
\subsection{Method \code{setup_prior()}}{
Set up the prior based on target data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$setup_prior(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data for the analysis}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorStep-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorStep-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorStep$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
