% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{generate_binary_data_from_rate}
\alias{generate_binary_data_from_rate}
\title{Generate binary data from a given rate}
\usage{
generate_binary_data_from_rate(rate, sample_size)
}
\arguments{
\item{rate}{The rate of success (1s) in the generated data}

\item{sample_size}{The total sample size}
}
\value{
A vector of binary data
}
\description{
This function generates binary data (0s and 1s) based on a given rate and sample size.
}
