% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_methods_comparison.R
\name{generate_comparison_table}
\alias{generate_comparison_table}
\title{Generate a comparison table}
\usage{
generate_comparison_table(
  results_df,
  x_metric,
  source_denominator_change_factor,
  target_to_source_std_ratio
)
}
\arguments{
\item{results_df}{A dataframe containing the results to be compared.}

\item{x_metric}{The metric to be used for comparison.}

\item{case_study}{The case study to filter the results.}

\item{target_sample_size_per_arm}{The target sample size per arm to filter the results.}
}
\value{
A kable object representing the comparison table.
}
\description{
This function generates a comparison table for different methods and treatment effects.
}
