% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{get_parameters}
\alias{get_parameters}
\title{Function to return a dataframe of parameters from json strings}
\usage{
get_parameters(parameters_df)
}
\arguments{
\item{parameters_df}{The dataframe containing parameters in JSON strings.}
}
\value{
An R dataframe with parsed parameters.
}
\description{
This function returns, for a dataframe containing parameters (in json strings) for a given method, an R dataframe.
}
\keyword{internal}
