% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{hellinger_distance}
\alias{hellinger_distance}
\title{Compute the Hellinger distance between two normal distributions.}
\usage{
hellinger_distance(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1}{Mean of the first normal distribution.}

\item{sigma1}{Standard deviation of the first normal distribution.}

\item{mu2}{Mean of the second normal distribution.}

\item{sigma2}{Standard deviation of the second normal distribution.}
}
\value{
The Hellinger distance between the two normal distributions.
}
\description{
Compute the Hellinger distance between two normal distributions.
}
