% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{important_drift_values}
\alias{important_drift_values}
\title{Get important drift values for a given source treatment effect and case study configuration.}
\usage{
important_drift_values(source_treatment_effect, case_study_config)
}
\arguments{
\item{source_treatment_effect}{The source treatment effect.}

\item{case_study_config}{Configuration for the case study.}
}
\value{
A vector of important drift values to consider.
}
\description{
Get important drift values for a given source treatment effect and case study configuration.
}
