% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_cache.R
\name{inference_cache_size}
\alias{inference_cache_size}
\title{Number of analyses held in the cache}
\usage{
inference_cache_size()
}
\value{
The number of distinct analyses stored.
}
\description{
Number of analyses held in the cache
}
