% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{make_labels_from_parameters}
\alias{make_labels_from_parameters}
\title{Function to make labels from parameters. Return a label formatted in Tex, for example "$\\xi_\\gamma$ = 0.5, $\\sigma_\\gamma$ = 0.1"}
\usage{
make_labels_from_parameters(parameter, method)
}
\arguments{
\item{parameter}{The parameter dataframe.}

\item{method}{The method name.}
}
\value{
The labels generated from the parameters.
}
\description{
Function to make labels from parameters. Return a label formatted in Tex, for example "$\\xi_\\gamma$ = 0.5, $\\sigma_\\gamma$ = 0.1"
}
\keyword{internal}
