% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{markers}
\alias{markers}
\title{Function to return markers from a list}
\usage{
markers(i, markers_list)
}
\arguments{
\item{i}{The index of the marker to return.}

\item{markers_list}{A list of markers.}
}
\value{
The marker at the specified index.
}
\description{
Function to return markers from a list
}
\keyword{internal}
