% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_RMP.R
\name{mixture_variance}
\alias{mixture_variance}
\title{Calculate the variance of a mixture distribution}
\usage{
mixture_variance(w, sigma2A, muA, sigma2B, muB)
}
\arguments{
\item{w}{Numeric. The weight of the first distribution in the mixture.}

\item{sigma2A}{Numeric. The variance of the first distribution.}

\item{muA}{Numeric. The mean of the first distribution.}

\item{sigma2B}{Numeric. The variance of the second distribution.}

\item{muB}{Numeric. The mean of the second distribution.}
}
\value{
Numeric. The total variance of the mixture distribution.
}
\description{
This function computes the variance of a mixture of two distributions.
}
\keyword{internal}
