% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{negative_binomial_regression}
\alias{negative_binomial_regression}
\title{Estimate rate and standard error}
\usage{
negative_binomial_regression(input_data)
}
\arguments{
\item{input_data}{The data for which the rate parameter and standard error need to be estimated.}
}
\value{
A list containing the rate estimate, standard error of the rate estimate, and standard error of the log(rate) estimate.
}
\description{
This function estimates the rate parameter of an intercept-only
negative binomial model and its standard error.
}
