% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_operating_characteristics_vs_tie.R
\name{operating_characteristic_vs_tie}
\alias{operating_characteristic_vs_tie}
\title{Function to generate an operating characteristic vs tie plot}
\usage{
operating_characteristic_vs_tie(
  results_metrics_df,
  case_study,
  target_sample_size_per_arm,
  treatment_effect,
  operating_characteristic,
  source_denominator_change_factor,
  target_to_source_std_ratio
)
}
\arguments{
\item{results_metrics_df}{The dataframe containing the results and metrics}

\item{case_study}{The case study name}

\item{target_sample_size_per_arm}{The target sample size per arm}

\item{treatment_effect}{The treatment effect type ("consistent", "no_effect", "partially_consistent")}

\item{operating_characteristic}{The operating characteristic to plot (e.g., "power", "type_1_error")}

\item{power_difference}{Logical indicating whether to calculate difference for the operating characteristic}
}
\value{
None
}
\description{
Function to generate an operating characteristic vs tie plot
}
