% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_operating_characteristics_vs_tie.R
\name{operating_characteristics_vs_tie_plots}
\alias{operating_characteristics_vs_tie_plots}
\title{Plot methods' operating characteristics}
\usage{
operating_characteristics_vs_tie_plots(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
OCs vs TIE
}
\examples{
NA
}
