% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_empirical_bayes_power_prior.R
\name{p_value_based_PP_Binomial}
\alias{p_value_based_PP_Binomial}
\title{Gaussian_empirical_Bayes_PP class}
\format{
R6Class object.
}
\description{
This is a parent class for variants of empirical Bayes PP methods for normally distributed summary measure of the treatment effect.
}
\section{Super classes}{
\code{\link[BExTE:Model]{Model}} -> \code{\link[BExTE:MCMCModel]{MCMCModel}} -> \code{\link[BExTE:BinomialCPP]{BinomialCPP}} -> \code{p_value_based_PP_Binomial}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{power_parameter}}{The power parameter.}

    \item{\code{summary_measure_likelihood}}{The summary measure distribution.}

    \item{\code{null_space}}{Null hypothesis space.}

    \item{\code{empirical_bayes}}{Boolean indicating if empirical Bayes is used.}

    \item{\code{shape_parameter}}{Shape parameter}

    \item{\code{equivalence_margin}}{Equivalence margin for the test}

    \item{\code{method}}{Method name}

    \item{\code{prior_var}}{Prior variance}

    \item{\code{mcmc_config}}{MCMC configuration}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-p_value_based_PP_Binomial-initialize}{\code{p_value_based_PP_Binomial$new()}}
    \item \href{#method-p_value_based_PP_Binomial-hypothesis_space_transformation}{\code{p_value_based_PP_Binomial$hypothesis_space_transformation()}}
    \item \href{#method-p_value_based_PP_Binomial-empirical_bayes_update}{\code{p_value_based_PP_Binomial$empirical_bayes_update()}}
    \item \href{#method-p_value_based_PP_Binomial-inference}{\code{p_value_based_PP_Binomial$inference()}}
    \item \href{#method-p_value_based_PP_Binomial-test}{\code{p_value_based_PP_Binomial$test()}}
    \item \href{#method-p_value_based_PP_Binomial-power_parameter_estimation}{\code{p_value_based_PP_Binomial$power_parameter_estimation()}}
    \item \href{#method-p_value_based_PP_Binomial-prior_pdf}{\code{p_value_based_PP_Binomial$prior_pdf()}}
    \item \href{#method-p_value_based_PP_Binomial-plot_power_parameter_vs_drift}{\code{p_value_based_PP_Binomial$plot_power_parameter_vs_drift()}}
    \item \href{#method-p_value_based_PP_Binomial-clone}{\code{p_value_based_PP_Binomial$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="check_data"><a href='../../BExTE/html/Model.html#method-Model-check_data'><code>Model$check_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="create"><a href='../../BExTE/html/Model.html#method-Model-create'><code>Model$create()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../BExTE/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../BExTE/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="inference_cache_scope"><a href='../../BExTE/html/Model.html#method-Model-inference_cache_scope'><code>Model$inference_cache_scope()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="plot_pdfs"><a href='../../BExTE/html/Model.html#method-Model-plot_pdfs'><code>Model$plot_pdfs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../BExTE/html/Model.html#method-Model-plot_posterior_pdf'><code>Model$plot_posterior_pdf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="plot_prior_pdf"><a href='../../BExTE/html/Model.html#method-Model-plot_prior_pdf'><code>Model$plot_prior_pdf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="posterior_beta_mixture"><a href='../../BExTE/html/Model.html#method-Model-posterior_beta_mixture'><code>Model$posterior_beta_mixture()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="posterior_mean"><a href='../../BExTE/html/Model.html#method-Model-posterior_mean'><code>Model$posterior_mean()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="posterior_moments"><a href='../../BExTE/html/Model.html#method-Model-posterior_moments'><code>Model$posterior_moments()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="posterior_quantile"><a href='../../BExTE/html/Model.html#method-Model-posterior_quantile'><code>Model$posterior_quantile()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="posterior_to_RBesT"><a href='../../BExTE/html/Model.html#method-Model-posterior_to_RBesT'><code>Model$posterior_to_RBesT()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="prior_ESS"><a href='../../BExTE/html/Model.html#method-Model-prior_ESS'><code>Model$prior_ESS()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="prior_elir_ess"><a href='../../BExTE/html/Model.html#method-Model-prior_elir_ess'><code>Model$prior_elir_ess()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="prior_to_RBesT"><a href='../../BExTE/html/Model.html#method-Model-prior_to_RBesT'><code>Model$prior_to_RBesT()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../BExTE/html/Model.html#method-Model-prior_treatment_benefit'><code>Model$prior_treatment_benefit()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../BExTE/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>Model$simulation_for_given_treatment_effect()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="test_decision"><a href='../../BExTE/html/Model.html#method-Model-test_decision'><code>Model$test_decision()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="Model" data-id="vectorised_replicate_inference"><a href='../../BExTE/html/Model.html#method-Model-vectorised_replicate_inference'><code>Model$vectorised_replicate_inference()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="check_mcmc_config"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-check_mcmc_config'><code>MCMCModel$check_mcmc_config()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="compute_posterior_parameters"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-compute_posterior_parameters'><code>MCMCModel$compute_posterior_parameters()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="credible_interval"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-credible_interval'><code>MCMCModel$credible_interval()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="posterior_cdf"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-posterior_cdf'><code>MCMCModel$posterior_cdf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="posterior_ess"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-posterior_ess'><code>MCMCModel$posterior_ess()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="posterior_median"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-posterior_median'><code>MCMCModel$posterior_median()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="posterior_pdf"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-posterior_pdf'><code>MCMCModel$posterior_pdf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="prior_cdf"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-prior_cdf'><code>MCMCModel$prior_cdf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="sample_posterior"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-sample_posterior'><code>MCMCModel$sample_posterior()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="MCMCModel" data-id="sample_prior"><a href='../../BExTE/html/MCMCModel.html#method-MCMCModel-sample_prior'><code>MCMCModel$sample_prior()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="BinomialCPP" data-id="draw_mcmc_prior"><a href='../../BExTE/html/BinomialCPP.html#method-BinomialCPP-draw_mcmc_prior'><code>BinomialCPP$draw_mcmc_prior()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="BExTE" data-topic="BinomialCPP" data-id="prepare_data"><a href='../../BExTE/html/BinomialCPP.html#method-BinomialCPP-prepare_data'><code>BinomialCPP$prepare_data()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-initialize}{}}}
\subsection{\code{p_value_based_PP_Binomial$new()}}{
  Initialize the p_value_based_PP object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$new(prior, theta_0, null_space, mcmc_config)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{prior}}{The prior object.}
      \item{\code{theta_0}}{Boundary of the null hypothesis space}
      \item{\code{null_space}}{Null space.}
      \item{\code{mcmc_config}}{MCMC configuration.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    None
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-hypothesis_space_transformation"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-hypothesis_space_transformation}{}}}
\subsection{\code{p_value_based_PP_Binomial$hypothesis_space_transformation()}}{
  Transform the hypothesis space.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$hypothesis_space_transformation(target_data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_data}}{The target data.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A list containing transformed treatment effect estimates.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-empirical_bayes_update"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-empirical_bayes_update}{}}}
\subsection{\code{p_value_based_PP_Binomial$empirical_bayes_update()}}{
  Empirical Bayes update
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$empirical_bayes_update(target_data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_data}}{Target study data}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL
Perform inference using the Gaussian_empirical_Bayes_PP method.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-inference"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-inference}{}}}
\subsection{\code{p_value_based_PP_Binomial$inference()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$inference(target_data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_data}}{The target data.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The inference result.
Test method
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-test"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-test}{}}}
\subsection{\code{p_value_based_PP_Binomial$test()}}{
  This method performs the test for the given target data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$test(
  target_data,
  source_treatment_effect_estimate,
  target_treatment_effect_estimate
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_data}}{The target data object.}
      \item{\code{source_treatment_effect_estimate}}{The source treatment effect estimate.}
      \item{\code{target_treatment_effect_estimate}}{The target treatment effect estimate.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The p-value.
Power parameter estimation method
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-power_parameter_estimation"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-power_parameter_estimation}{}}}
\subsection{\code{p_value_based_PP_Binomial$power_parameter_estimation()}}{
  This method estimates the power parameter for the given target data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$power_parameter_estimation(target_data)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_data}}{The target data object.}
      \item{\code{source_treatment_effect_estimate}}{The source treatment effect estimate.}
      \item{\code{target_treatment_effect_estimate}}{The target treatment effect estimate.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The power parameter.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-prior_pdf}{}}}
\subsection{\code{p_value_based_PP_Binomial$prior_pdf()}}{
  Calculate the prior probability density function (PDF) for a given target treatment effect.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$prior_pdf(target_treatment_effect)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{target_treatment_effect}}{The target treatment effect.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The prior PDF.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-plot_power_parameter_vs_drift"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-plot_power_parameter_vs_drift}{}}}
\subsection{\code{p_value_based_PP_Binomial$plot_power_parameter_vs_drift()}}{
  Plot the power parameter as a function of drift in treatment effect
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$plot_power_parameter_vs_drift(
  source_treatment_effect_estimate,
  target_data,
  min_drift,
  max_drift,
  resolution
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{source_treatment_effect_estimate}}{Treatment effect estimate in the source study}
      \item{\code{target_data}}{Target study data}
      \item{\code{min_drift}}{Minimum drift value}
      \item{\code{max_drift}}{Maximum drift value}
      \item{\code{resolution}}{Number of points on the drift grid.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The prior PDF.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP_Binomial-clone"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP_Binomial-clone}{}}}
\subsection{\code{p_value_based_PP_Binomial$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{p_value_based_PP_Binomial$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
