% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_empirical_bayes_hyperparameters_vs_scenario.R
\name{plot_empirical_bayes_hyperparameters_vs_drift}
\alias{plot_empirical_bayes_hyperparameters_vs_drift}
\title{Plot Empirical Bayes Hyperparameters vs Drift}
\usage{
plot_empirical_bayes_hyperparameters_vs_drift(
  results_metrics_df,
  method,
  case_study,
  control_drift,
  xvars,
  source_denominator_change_factor,
  target_to_source_std_ratio,
  parameters_combinations
)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{method}{The method to plot the parameters for.}

\item{case_study}{The case study to plot the parameters for.}

\item{control_drift}{A logical value indicating whether to filter the data based on control drift.}

\item{xvars}{Variable on the x axis}
}
\value{
None
}
\description{
This function plots the hyperparameters updated using empirical Bayes against the drift for a given method, case study sample size, and control drift option.
}
