% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_operating_characteristics_vs_drift_all_methods.R
\name{plot_metric_vs_drift_methods}
\alias{plot_metric_vs_drift_methods}
\title{Function to generate a metric vs drift plot}
\usage{
plot_metric_vs_drift_methods(
  metric,
  results_metrics_df,
  theta_0,
  case_study,
  target_sample_size_per_arm,
  xvars,
  target_to_source_std_ratio = 1,
  join_points = FALSE,
  source_denominator_change_factor = 1
)
}
\arguments{
\item{metric}{The metric to plot}

\item{results_metrics_df}{The dataframe containing the results and metrics}

\item{theta_0}{The true treatment effect}

\item{case_study}{The case study name}

\item{target_sample_size_per_arm}{The target sample size per arm}

\item{xvars}{A list of x-variables for control drift and treatment drift}

\item{target_to_source_std_ratio}{Ratio between the target and source studies sampling standard deviation.}

\item{join_points}{Whether to join the point with a line or not.}

\item{method}{The method name}

\item{category}{The category to group the results by (either "parameters" or "target_sample_size_per_arm")}

\item{parameters_combinations}{The combinations of parameters to filter the results by}
}
\value{
None
}
\description{
Function to generate a metric vs drift plot
}
