% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequentist_operating_characteristics.R
\name{plot_metric_vs_sample_size}
\alias{plot_metric_vs_sample_size}
\title{Function to plot metric vs sample size}
\usage{
plot_metric_vs_sample_size(
  metric,
  results_metrics_df,
  case_study,
  method,
  parameters_combinations,
  theta_0 = 0,
  add_baselines = FALSE,
  target_to_source_std_ratio = 1,
  source_denominator_change_factor = 1,
  analysis_config
)
}
\arguments{
\item{metric}{The metric to plot.}

\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{case_study}{The case study to plot the metric for.}

\item{method}{The method to plot the metric for.}

\item{parameters_combinations}{The parameter combinations to filter the data on.}

\item{theta_0}{Boundary of the null hypothesis space.}

\item{add_baselines}{A logical value indicating whether to add baselines to the plot.}
}
\value{
None
}
\description{
This function plots a metric against the sample size for a given method, case study, and parameters combinations.
}
