% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_operating_characteristics_vs_drift_all_methods.R
\name{plot_metric_vs_sample_size_methods}
\alias{plot_metric_vs_sample_size_methods}
\title{Function to plot metric vs sample size}
\usage{
plot_metric_vs_sample_size_methods(
  metric,
  results_metrics_df,
  case_study,
  theta_0 = 0,
  target_to_source_std_ratio = 1,
  source_denominator_change_factor = 1,
  dodging = FALSE,
  join_points = FALSE,
  target_treatment_effect = target_treatment_effect
)
}
\arguments{
\item{metric}{The metric to plot.}

\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{case_study}{The case study to plot the metric for.}

\item{theta_0}{Boundary of the null hypothesis space.}
}
\value{
None
}
\description{
This function plots a metric against the sample size for a given method, case study, and parameters combinations.
}
