% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_operating_characteristics_vs_drift_all_methods.R
\name{plot_metric_vs_scenario_methods}
\alias{plot_metric_vs_scenario_methods}
\title{Plot methods operating characteristics}
\usage{
plot_metric_vs_scenario_methods(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of different methods.
}
\examples{
NA
}
