% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_vs_ess.R
\name{plot_metrics_vs_ess}
\alias{plot_metrics_vs_ess}
\title{Plot methods operating characteristics}
\usage{
plot_metrics_vs_ess(results_metrics_df, frequentist_metrics, inference_metrics)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of different methods.
}
\examples{
NA
}
