% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_success_probability_vs_scenario.R
\name{plot_success_proba_vs_drift}
\alias{plot_success_proba_vs_drift}
\title{Function to generate a metric vs drift plot}
\usage{
plot_success_proba_vs_drift(
  metric,
  results_metrics_df,
  theta_0,
  case_study,
  method,
  target_sample_size_per_arm = NULL,
  target_to_source_std_ratio,
  source_denominator_change_factor = NULL,
  parameters_combinations = NULL,
  xvars,
  join_points = FALSE,
  baseline_success_proba
)
}
\arguments{
\item{metric}{The metric to plot}

\item{results_metrics_df}{The dataframe containing the results and metrics}

\item{theta_0}{The true treatment effect}

\item{case_study}{The case study name}

\item{method}{The method name}

\item{target_sample_size_per_arm}{The target sample size per arm}

\item{target_to_source_std_ratio}{Ratio between the target and source studies sampling standard deviation.}

\item{parameters_combinations}{The combinations of parameters to filter the results by}

\item{xvars}{A list of x-variables for control drift and treatment drift}

\item{join_points}{Whether to join the point with a line or not.}
}
\value{
None
}
\description{
Function to generate a metric vs drift plot
}
