% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sweet_spots_comparison.R
\name{plot_sweet_spot_width_vs_scenario}
\alias{plot_sweet_spot_width_vs_scenario}
\title{Plot methods operating characteristics}
\usage{
plot_sweet_spot_width_vs_scenario(sweet_spot_df, sweet_spots_metrics)
}
\arguments{
\item{sweet_spot_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of different methods.
}
\examples{
NA
}
