% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_parameters_vs_scenario.R
\name{posterior_parameters_plots}
\alias{posterior_parameters_plots}
\title{Plot frequentist methods operating characteristics}
\usage{
posterior_parameters_plots(results_metrics_df)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of frequentist methods.
}
\examples{
NA
}
