% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_posterior_parameters_vs_scenario.R
\name{posterior_parameters_tables}
\alias{posterior_parameters_tables}
\title{Generate multiple posterior parameter tables}
\usage{
posterior_parameters_tables(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{A data frame containing the results metrics.}

\item{metrics}{A list or vector of metrics to include in the analysis.}
}
\value{
This function does not return a value. It calls other functions to create and save multiple tables.
}
\description{
This function generates multiple tables for posterior parameters across different case studies, methods, and sample sizes.
}
