% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_vs_tie.R
\name{power_vs_tie_plots}
\alias{power_vs_tie_plots}
\title{Plot frequentist methods operating characteristics}
\usage{
power_vs_tie_plots(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The data frame containing the results and metrics.}

\item{metrics}{A list of metrics to be plotted.}
}
\value{
None
}
\description{
This function generates plots for the operating characteristics of frequentist methods.
}
\examples{
NA
}
