% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_bayesian_ocs.R
\name{preposterior_proba_FP_MC}
\alias{preposterior_proba_FP_MC}
\title{Calculate the probability of false positive}
\usage{
preposterior_proba_FP_MC(
  conditional_proba_success,
  design_prior_samples,
  theta_0,
  null_space
)
}
\arguments{
\item{conditional_proba_success}{The conditional probability of success.}

\item{design_prior_samples}{The design prior samples.}

\item{theta_0}{The value of theta_0.}

\item{null_space}{The null space (either "left" or "right").}
}
\value{
The probability of false positive.
}
\description{
This function calculates the probability of false positive based on the conditional probability of success,
design prior samples, theta_0, and null space.
}
\examples{
NA
}
