% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_prior_ess_estimation.R
\name{prior_ess_elir}
\alias{prior_ess_elir}
\title{Calculate the priorEffective Sample Size (ESS) based on ELIR for a Bayesian model.}
\usage{
prior_ess_elir(rbest_model, target_data)
}
\arguments{
\item{rbest_model}{A RBesT model object.}

\item{target_data}{Target data containing sample information.}
}
\value{
The prior moment-based ESS for the Bayesian model.
}
\description{
This function calculates the ELIR for a Bayesian model using the RBesT package.
It takes a RBesT model object and target data as input and returns the prior moment-based ESS.
}
