% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_prior_ess_estimation.R
\name{prior_precision_ess}
\alias{prior_precision_ess}
\title{Calculate the prior precision-based Effective Sample Size (ESS) for a Bayesian model.}
\usage{
prior_precision_ess(rbest_model, target_data)
}
\arguments{
\item{rbest_model}{A RBesT model object.}

\item{target_data}{Target data containing sample information.}
}
\value{
The prior precision-based ESS for the Bayesian model.
}
\description{
This function calculates the prior precision-based ESS for a Bayesian model using the RBesT package.
It takes a RBesT model object and target data as input and returns the prior precision-based ESS.
}
