% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{rate_from_drift_logOR}
\alias{rate_from_drift_logOR}
\title{Calculate the success rate in the target study arm from the drift on the log odds ratio scale}
\usage{
rate_from_drift_logOR(arm_drift, source_rate)
}
\arguments{
\item{arm_drift}{The drift (defined on the log scale) between the target and source study arms}

\item{source_rate}{The success rate in the arm of interest in the source study}
}
\value{
The success rate in the corresponding arm of the target study
}
\description{
This function calculates the success rate in the target study arm based on the drift (defined on the log scale) and the success rate in the arm of interest in the source study.
}
\keyword{internal}
