% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBexT_package.R
\docType{package}
\name{RBExT-package}
\alias{RBExT}
\alias{RBExT-package}
\title{RBExT: Bayesian extrapolation with R}
\description{
This package provides tools for Bayesian extrapolation with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/quintengroup/EMA_ROC02_Bayesian_simulation}
  \item Report bugs at \url{https://github.com/quintengroup/EMA_ROC02_Bayesian_simulation/issues}
}

}
\author{
\strong{Maintainer}: Tristan Fauvel \email{t.fauvel@quinten-health.com}

Authors:
\itemize{
  \item Pascal Godbillot
}

}
\keyword{internal}
