% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_columns_from_df}
\alias{remove_columns_from_df}
\title{Remove columns from a dataframe}
\usage{
remove_columns_from_df(df, to_remove)
}
\arguments{
\item{df}{A dataframe to modify.}

\item{to_remove}{A character vector of column names to remove.}
}
\value{
A dataframe with specified columns removed.
}
\description{
This function removes specified columns from a dataframe.
}
\keyword{internal}
