% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{sample_aggregate_normal_data}
\alias{sample_aggregate_normal_data}
\title{Sample aggregate normal data based on the mean and variance}
\usage{
sample_aggregate_normal_data(mean, variance, n_replicates, n_samples_per_arm)
}
\arguments{
\item{mean}{The mean of the normal distribution}

\item{variance}{The variance of the normal distribution}

\item{n_replicates}{Number of replicates to sample}

\item{n_samples_per_arm}{Number of samples per replicate}
}
\value{
A list containing the sample mean and sample standard error for each replicate
}
\description{
This function samples aggregate normal data based on the mean and variance.
}
