% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{sample_log_odds_ratios}
\alias{sample_log_odds_ratios}
\title{Sample log odds ratios based on the number of participants and responders in each arm}
\usage{
sample_log_odds_ratios(
  n_control,
  n_treatment,
  treatment_rate,
  control_rate,
  n_replicates
)
}
\arguments{
\item{n_control}{Number of participants in the control arm}

\item{n_treatment}{Number of participants in the treatment arm}

\item{treatment_rate}{The rate of success (1s) in the treatment arm}

\item{control_rate}{The rate of success (1s) in the control arm}

\item{n_replicates}{Number of replicates to sample}
}
\value{
A list containing the sampled log odds ratios and the standard error of the log odds ratio
}
\description{
This function samples log odds ratios based on the number of participants and responders in the control and treatment arms.
}
