% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{sample_rate_ratios}
\alias{sample_rate_ratios}
\title{Sample rate ratios based on the rates in the control and treatment arms}
\usage{
sample_rate_ratios(
  control_rate,
  treatment_rate,
  n_replicates,
  n_control,
  n_treatment
)
}
\arguments{
\item{control_rate}{The rate in the control arm}

\item{treatment_rate}{The rate in the treatment arm}

\item{n_replicates}{Number of replicates to sample}

\item{n_control}{Number of participants in the control arm}

\item{n_treatment}{Number of participants in the treatment arm}
}
\value{
A vector of sampled rate ratios
}
\description{
This function samples rate ratios based on the rates in the control and treatment arms.
}
