% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validation.R
\name{scenarios_config_schema}
\alias{scenarios_config_schema}
\title{Schema for \code{scenarios_config.yml}}
\usage{
scenarios_config_schema
}
\description{
Schema for \code{scenarios_config.yml}
}
\keyword{internal}
