% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_utils.R
\name{set_size}
\alias{set_size}
\title{Function to set figure dimensions}
\usage{
set_size(width, fraction = 1, aspect_ratio = NULL)
}
\arguments{
\item{width}{The desired width of the figure (in pts).}

\item{fraction}{The fraction of the width to use (default is 1).}

\item{aspect_ratio}{The desired aspect ratio of the figure (default is golden ratio).}
}
\value{
A vector containing the width and height of the figure.
}
\description{
Function to set figure dimensions
}
\keyword{internal}
