% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validation.R
\name{simulation_config_schema}
\alias{simulation_config_schema}
\title{Schema for \code{simulation_config.yml}}
\usage{
simulation_config_schema
}
\description{
Schema for \code{simulation_config.yml}
}
\keyword{internal}
