% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{standard_error_log_odds_ratio}
\alias{standard_error_log_odds_ratio}
\title{Compute the standard error of the log odds ratio}
\usage{
standard_error_log_odds_ratio(
  n_control_nonresponders,
  n_treatment_nonresponders,
  n_control_responders,
  n_treatment_responders,
  continuity_correction = TRUE
)
}
\arguments{
\item{n_control_nonresponders}{Number of noncases in the unexposed group}

\item{n_treatment_nonresponders}{Number of noncases in the exposed group}

\item{n_control_responders}{Number of cases in the unexposed group}

\item{n_treatment_responders}{Number of cases in the exposed group}

\item{continuity_correction}{Logical; if TRUE, applies continuity correction (default is TRUE)}
}
\value{
The standard error of the log odds ratio

The standard error of the log odds ratio
}
\description{
This function calculates the standard error of the log odds ratio based on the number of noncases and cases in the unexposed and exposed groups.
}
\examples{
standard_error_log_odds_ratio(100, 90, 50, 60)

}
\keyword{internal}
