% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_compute_sweet_spot.R
\name{sweet_spot_determination}
\alias{sweet_spot_determination}
\title{Determine Sweet Spot Bounds and Width}
\usage{
sweet_spot_determination(
  x_values,
  y_values,
  reference_value = 0,
  larger_is_better,
  return_NA_at_boundaries = FALSE
)
}
\arguments{
\item{x_values}{A numeric vector of x values.}

\item{y_values}{A numeric vector of y values.}

\item{reference_value}{A numeric reference value to determine the crossing points.}
}
\value{
A list containing:
\itemize{
\item sweet_spot_bounds A numeric vector of length 2 with the lower and upper bounds of the sweet spot.
\item sweet_spot_widthA numeric value indicating the width of the sweet spot.
}
}
\description{
This function calculates the bounds and width of the "sweet spot" where a given \code{y_values} vector
crosses a specified \code{reference_value}. It uses linear interpolation for precise crossing points.
}
