% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bayesian_operating_characteristics.R
\name{table_bayesian_metrics}
\alias{table_bayesian_metrics}
\title{Generate tables for Bayesian metrics}
\usage{
table_bayesian_metrics(
  results_metrics_df,
  case_study,
  target_sample_size_per_arm,
  source_denominator_change_factor = 1,
  target_to_source_std_ratio = 1
)
}
\arguments{
\item{results_metrics_df}{A data frame containing the results and metrics.}

\item{case_study}{Character string specifying the case study.}
}
\value{
This function doesn't return a value but saves the generated tables as HTML, PDF, and LaTeX files.
}
\description{
Generate tables for Bayesian metrics
}
