% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bayesian_operating_characteristics.R
\name{table_bayesian_ocs}
\alias{table_bayesian_ocs}
\title{Generate tables for methods operating characteristics}
\usage{
table_bayesian_ocs(
  results_metrics_df,
  metrics,
  source_denominator_change_factor = 1,
  target_to_source_std_ratio = 1
)
}
\arguments{
\item{results_metrics_df}{A data frame containing the results and metrics.}

\item{metrics}{A vector of metric names to include in the tables.}
}
\value{
This function doesn't return a value but calls other functions to generate and save tables.
}
\description{
This function creates tables for various Bayesian operating characteristics across different case studies and methods.
}
