% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_case_study_summary.R
\name{table_case_study_summary}
\alias{table_case_study_summary}
\title{Summary of the clinical case studies (table S7)}
\usage{
table_case_study_summary(case_studies, case_studies_config_dir, tables_dir)
}
\arguments{
\item{case_studies}{Character vector of case study names.}

\item{case_studies_config_dir}{Directory holding the case study YAMLs.}

\item{tables_dir}{Directory to write into.}
}
\value{
The path of the written \code{.tex} file.
}
\description{
Summary of the clinical case studies (table S7)
}
