% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/table_empirical_bayes_hyperparameters_vs_scenario.R
\name{table_empirical_bayes_hyperparameters_vs_scenario}
\alias{table_empirical_bayes_hyperparameters_vs_scenario}
\title{This function tables for hyperparameters updated using Empirical Bayes across different case studies, methods, and sample sizes.}
\usage{
table_empirical_bayes_hyperparameters_vs_scenario(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{A data frame containing the results metrics.}

\item{metrics}{A list or vector of metrics to include in the analysis.}
}
\value{
This function does not return a value. It calls other functions to create and save multiple tables.
}
\description{
This function tables for hyperparameters updated using Empirical Bayes across different case studies, methods, and sample sizes.
}
