% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_methods_comparison.R
\name{table_methods_comparison}
\alias{table_methods_comparison}
\title{Methods comparison in a table}
\usage{
table_methods_comparison(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{The results metrics dataframe.}

\item{metrics}{List of frequentist metrics and their properties.}
}
\value{
None
}
\description{
This function creates tables comparing methods with respect to different metrics for various case studies and sample sizes.
}
