% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_frequentist_operating_characteristics.R
\name{table_metric_vs_drift}
\alias{table_metric_vs_drift}
\title{Generate a metric vs drift table}
\usage{
table_metric_vs_drift(
  metric,
  results_metrics_df,
  theta_0,
  case_study,
  method,
  category,
  control_drift,
  target_sample_size_per_arm,
  parameters_combinations,
  xvars,
  source_denominator_change_factor = 1,
  target_to_source_std_ratio = 1,
  wide_table = TRUE
)
}
\arguments{
\item{metric}{The metric to be plotted}

\item{results_metrics_df}{A dataframe containing the results metrics}

\item{theta_0}{The null hypothesis value}

\item{case_study}{The case study being analyzed}

\item{method}{The method being used}

\item{category}{The category of analysis ('parameters' or 'target_sample_size_per_arm')}

\item{control_drift}{Boolean indicating whether to control for drift}

\item{target_sample_size_per_arm}{The target sample size per arm}

\item{parameters_combinations}{The combinations of parameters}

\item{xvars}{A list containing x-axis variable information}
}
\value{
Generates and saves tables in HTML, PDF, and LaTeX formats
}
\description{
Generate a metric vs drift table
}
