% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_frequentist_operating_characteristics.R
\name{table_metric_vs_sample_size}
\alias{table_metric_vs_sample_size}
\title{Plot metric vs sample size}
\usage{
table_metric_vs_sample_size(
  metric,
  results_metrics_df,
  case_study,
  method,
  theta_0 = 0,
  source_denominator_change_factor = 1,
  target_to_source_std_ratio = 1,
  wide_table = TRUE
)
}
\arguments{
\item{metric}{The metric to be plotted}

\item{results_metrics_df}{A dataframe containing the results metrics}

\item{case_study}{The case study being analyzed}

\item{method}{The method being used}

\item{theta_0}{The null hypothesis value}

\item{parameters_combinations}{The combinations of parameters}
}
\value{
Generates and saves tables in HTML, PDF, and LaTeX formats
}
\description{
Plot metric vs sample size
}
