% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_frequentist_operating_characteristics.R
\name{tables_metric_vs_scenario}
\alias{tables_metric_vs_scenario}
\title{Generate tables for methods operating characteristics}
\usage{
tables_metric_vs_scenario(results_metrics_df, metrics)
}
\arguments{
\item{results_metrics_df}{A dataframe containing the results metrics}

\item{metrics}{A list of metrics to be analyzed}
}
\value{
Generates and saves multiple tables for different metrics and methods
}
\description{
Generate tables for methods operating characteristics
}
