% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorised_frequentist_tests.R
\name{tail_probabilities}
\alias{tail_probabilities}
\title{Convert test statistics to p-values for a given alternative}
\usage{
tail_probabilities(statistic, alternative, distribution)
}
\arguments{
\item{statistic}{Vector of test statistics.}

\item{alternative}{One of \code{"two.sided"}, \code{"less"} or \code{"greater"}.}

\item{distribution}{Function of \verb{(q, lower)} giving the null distribution.}
}
\value{
Vector of p-values.
}
\description{
Convert test statistics to p-values for a given alternative
}
\keyword{internal}
