% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_design_prior.R
\name{AnalysisPriorDesignPrior}
\alias{AnalysisPriorDesignPrior}
\title{AnalysisPriorDesignPrior class}
\description{
A class representing the analysis prior design prior for Bayesian borrowing.
}
\section{Super class}{
\code{\link[RBExT:DesignPrior]{RBExT::DesignPrior}} -> \code{AnalysisPriorDesignPrior}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Model used to define the analysis prior}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalysisPriorDesignPrior-new}{\code{AnalysisPriorDesignPrior$new()}}
\item \href{#method-AnalysisPriorDesignPrior-sample}{\code{AnalysisPriorDesignPrior$sample()}}
\item \href{#method-AnalysisPriorDesignPrior-cdf}{\code{AnalysisPriorDesignPrior$cdf()}}
\item \href{#method-AnalysisPriorDesignPrior-pdf}{\code{AnalysisPriorDesignPrior$pdf()}}
\item \href{#method-AnalysisPriorDesignPrior-clone}{\code{AnalysisPriorDesignPrior$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="DesignPrior" data-id="create"><a href='../../RBExT/html/DesignPrior.html#method-DesignPrior-create'><code>RBExT::DesignPrior$create()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisPriorDesignPrior-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisPriorDesignPrior-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new instance of AnalysisPriorDesignPrior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisPriorDesignPrior$new(
  model,
  case_study_config,
  simulation_config,
  mcmc_config
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The model object.}

\item{\code{case_study_config}}{Case study configuration}

\item{\code{simulation_config}}{Simulation configuration}

\item{\code{mcmc_config}}{MCMC configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisPriorDesignPrior-sample"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisPriorDesignPrior-sample}{}}}
\subsection{Method \code{sample()}}{
Samples from the analysis prior design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisPriorDesignPrior$sample(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisPriorDesignPrior-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisPriorDesignPrior-cdf}{}}}
\subsection{Method \code{cdf()}}{
Computes the cumulative distribution function (CDF) of the analysis prior design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisPriorDesignPrior$cdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the CDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisPriorDesignPrior-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisPriorDesignPrior-pdf}{}}}
\subsection{Method \code{pdf()}}{
Computes the PDF of the analysis prior design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisPriorDesignPrior$pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the PDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalysisPriorDesignPrior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalysisPriorDesignPrior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalysisPriorDesignPrior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
