% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_power_priors.R
\name{BinomialCPP}
\alias{BinomialCPP}
\title{BinomialCPP Class}
\description{
A class representing a Binomial Conditional Power Prior model.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:MCMCModel]{RBExT::MCMCModel}} -> \code{BinomialCPP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{power_parameter}}{The power parameter for the model}

\item{\code{method}}{Method name}

\item{\code{stan_prior}}{Stan prior model}

\item{\code{stan_prior_code}}{Stan code for the prior}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinomialCPP-new}{\code{BinomialCPP$new()}}
\item \href{#method-BinomialCPP-prepare_data}{\code{BinomialCPP$prepare_data()}}
\item \href{#method-BinomialCPP-draw_mcmc_prior}{\code{BinomialCPP$draw_mcmc_prior()}}
\item \href{#method-BinomialCPP-clone}{\code{BinomialCPP$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="empirical_bayes_update"><a href='../../RBExT/html/Model.html#method-Model-empirical_bayes_update'><code>RBExT::Model$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_mean"><a href='../../RBExT/html/Model.html#method-Model-posterior_mean'><code>RBExT::Model$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_moments"><a href='../../RBExT/html/Model.html#method-Model-posterior_moments'><code>RBExT::Model$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-posterior_to_RBesT'><code>RBExT::Model$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-prior_to_RBesT'><code>RBExT::Model$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="check_data"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-check_data'><code>RBExT::MCMCModel$check_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="check_mcmc_config"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-check_mcmc_config'><code>RBExT::MCMCModel$check_mcmc_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="compute_posterior_parameters"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-compute_posterior_parameters'><code>RBExT::MCMCModel$compute_posterior_parameters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="credible_interval"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-credible_interval'><code>RBExT::MCMCModel$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="inference"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-inference'><code>RBExT::MCMCModel$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_cdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_cdf'><code>RBExT::MCMCModel$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_median"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_median'><code>RBExT::MCMCModel$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_pdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_pdf'><code>RBExT::MCMCModel$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="prior_cdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-prior_cdf'><code>RBExT::MCMCModel$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="prior_pdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-prior_pdf'><code>RBExT::MCMCModel$prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="sample_posterior"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-sample_posterior'><code>RBExT::MCMCModel$sample_posterior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="sample_prior"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-sample_prior'><code>RBExT::MCMCModel$sample_prior()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinomialCPP-new"></a>}}
\if{latex}{\out{\hypertarget{method-BinomialCPP-new}{}}}
\subsection{Method \code{new()}}{
Initialize an instance of the BinomialCPP class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialCPP$new(prior, mcmc_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior object}

\item{\code{mcmc_config}}{The MCMC configuration parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinomialCPP-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-BinomialCPP-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
Prepare data for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialCPP$prepare_data(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target study data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of prepared data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinomialCPP-draw_mcmc_prior"></a>}}
\if{latex}{\out{\hypertarget{method-BinomialCPP-draw_mcmc_prior}{}}}
\subsection{Method \code{draw_mcmc_prior()}}{
Sample from the prior using Stan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialCPP$draw_mcmc_prior()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinomialCPP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinomialCPP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialCPP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
