% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_colnames}
\alias{check_colnames}
\title{Check columns in a dataframe}
\usage{
check_colnames(df, expected_colnames)
}
\arguments{
\item{df}{A dataframe to check.}

\item{expected_colnames}{A character vector of column names the dataframe should contain.}
}
\value{
No return value, called for side effects.
}
\description{
This function checks if a dataframe contains the expected columns and only the expected columns.
}
\keyword{internal}
