% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_design_prior.R
\name{DesignPrior}
\alias{DesignPrior}
\title{DesignPrior Class}
\description{
A class representing the design prior for Bayesian borrowing.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{design_prior_type}}{Type of design prior.}

\item{\code{RBesT_model}}{RBesT version of the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignPrior-create}{\code{DesignPrior$create()}}
\item \href{#method-DesignPrior-sample}{\code{DesignPrior$sample()}}
\item \href{#method-DesignPrior-cdf}{\code{DesignPrior$cdf()}}
\item \href{#method-DesignPrior-pdf}{\code{DesignPrior$pdf()}}
\item \href{#method-DesignPrior-clone}{\code{DesignPrior$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignPrior-create"></a>}}
\if{latex}{\out{\hypertarget{method-DesignPrior-create}{}}}
\subsection{Method \code{create()}}{
Creates a new instance of DesignPrior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignPrior$create(
  design_prior_type,
  model,
  source_data,
  case_study_config,
  simulation_config,
  mcmc_config,
  case_study
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design_prior_type}}{The type of design prior.}

\item{\code{model}}{The model object.}

\item{\code{source_data}}{The source data object.}

\item{\code{case_study_config}}{Case study configuration}

\item{\code{simulation_config}}{Simulation configuration}

\item{\code{mcmc_config}}{MCMC configuration}

\item{\code{case_study}}{Case study name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of DesignPrior.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignPrior-sample"></a>}}
\if{latex}{\out{\hypertarget{method-DesignPrior-sample}{}}}
\subsection{Method \code{sample()}}{
Samples from the design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignPrior$sample(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignPrior-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-DesignPrior-cdf}{}}}
\subsection{Method \code{cdf()}}{
Computes the cumulative distribution function (CDF) of the design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignPrior$cdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the CDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignPrior-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-DesignPrior-pdf}{}}}
\subsection{Method \code{pdf()}}{
Computes the PDF of the design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignPrior$pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the PDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignPrior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignPrior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignPrior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
