% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_frequentist_ocs.R
\name{frequentist_ocs_scenario_simulation}
\alias{frequentist_ocs_scenario_simulation}
\title{Simulate a Scenario}
\usage{
frequentist_ocs_scenario_simulation(
  scenario,
  simulation_config,
  scenarios_config,
  freq_filename,
  config_dir,
  case_studies_config_dir
)
}
\arguments{
\item{scenario}{The scenario to simulate.}

\item{simulation_config}{Simulation configuration}

\item{freq_filename}{File name for the frequentist OCs}

\item{config_dir}{Configuration files directory}

\item{case_studies_config_dir}{Case studies configurations directory}
}
\value{
The results of the simulation.
}
\description{
This function simulates a scenario based on the provided inputs and returns the results of the simulation.
}
