% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_power_priors.R
\name{Gaussian_NPP}
\alias{Gaussian_NPP}
\title{Gaussian_NPP class}
\description{
This class represents a Gaussian model using the NPP (Noninformative Power Prior) approach.
It inherits from the Model class.
}
\section{Super class}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{Gaussian_NPP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{power_parameter_mean}}{The mean of the prior on the power parameter.}

\item{\code{power_parameter_std}}{The standard deviation of the prior on the power parameter.}

\item{\code{target_treatment_effect_estimate}}{Estimate of the treatment effect in the target study}

\item{\code{target_treatment_effect_standard_error}}{Standard error on the estimate of the treatment effect in the target study}

\item{\code{p}}{Parameter of the initial Beta prior on the power parameter}

\item{\code{q}}{Parameter of the initial Beta prior on the power parameter}

\item{\code{prior_normconst}}{Normalization constant of the prior}

\item{\code{posterior_normconst}}{Normalization constant of the posterior}

\item{\code{max_posterior_pdf}}{Maximum value of the posterior p.d.f., used for rejection sampling of the posterior p.d.f.}

\item{\code{method}}{Name of the method}

\item{\code{summary_measure_likelihood}}{Summary measure likelihood}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Gaussian_NPP-new}{\code{Gaussian_NPP$new()}}
\item \href{#method-Gaussian_NPP-unnormalized_posterior_power_parameter_pdf}{\code{Gaussian_NPP$unnormalized_posterior_power_parameter_pdf()}}
\item \href{#method-Gaussian_NPP-power_parameter_posterior_pdf}{\code{Gaussian_NPP$power_parameter_posterior_pdf()}}
\item \href{#method-Gaussian_NPP-normalizing_constant_power_parameter}{\code{Gaussian_NPP$normalizing_constant_power_parameter()}}
\item \href{#method-Gaussian_NPP-inference}{\code{Gaussian_NPP$inference()}}
\item \href{#method-Gaussian_NPP-credible_interval}{\code{Gaussian_NPP$credible_interval()}}
\item \href{#method-Gaussian_NPP-posterior_median}{\code{Gaussian_NPP$posterior_median()}}
\item \href{#method-Gaussian_NPP-sample_posterior}{\code{Gaussian_NPP$sample_posterior()}}
\item \href{#method-Gaussian_NPP-sample_prior}{\code{Gaussian_NPP$sample_prior()}}
\item \href{#method-Gaussian_NPP-posterior_cdf}{\code{Gaussian_NPP$posterior_cdf()}}
\item \href{#method-Gaussian_NPP-posterior_pdf}{\code{Gaussian_NPP$posterior_pdf()}}
\item \href{#method-Gaussian_NPP-prior_pdf}{\code{Gaussian_NPP$prior_pdf()}}
\item \href{#method-Gaussian_NPP-plot_power_parameter_posterior_pdf}{\code{Gaussian_NPP$plot_power_parameter_posterior_pdf()}}
\item \href{#method-Gaussian_NPP-plot_power_parameter_vs_drift}{\code{Gaussian_NPP$plot_power_parameter_vs_drift()}}
\item \href{#method-Gaussian_NPP-clone}{\code{Gaussian_NPP$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="empirical_bayes_update"><a href='../../RBExT/html/Model.html#method-Model-empirical_bayes_update'><code>RBExT::Model$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_mean"><a href='../../RBExT/html/Model.html#method-Model-posterior_mean'><code>RBExT::Model$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_moments"><a href='../../RBExT/html/Model.html#method-Model-posterior_moments'><code>RBExT::Model$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-posterior_to_RBesT'><code>RBExT::Model$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_cdf"><a href='../../RBExT/html/Model.html#method-Model-prior_cdf'><code>RBExT::Model$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-prior_to_RBesT'><code>RBExT::Model$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-new"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Gaussian_NPP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$new(prior)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{Prior object containing method parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Gaussian_NPP object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-unnormalized_posterior_power_parameter_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-unnormalized_posterior_power_parameter_pdf}{}}}
\subsection{Method \code{unnormalized_posterior_power_parameter_pdf()}}{
Calculate the unnormalized posterior power parameter PDF.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$unnormalized_posterior_power_parameter_pdf(
  power_parameter,
  target_data
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{power_parameter}}{Power parameter value.}

\item{\code{target_data}}{Target study data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Unnormalized posterior power parameter PDF value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-power_parameter_posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-power_parameter_posterior_pdf}{}}}
\subsection{Method \code{power_parameter_posterior_pdf()}}{
Return the posterior distribution of the power parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$power_parameter_posterior_pdf(power_parameter, target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{power_parameter}}{Power parameter value.}

\item{\code{target_data}}{Target study data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Posterior power parameter PDF value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-normalizing_constant_power_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-normalizing_constant_power_parameter}{}}}
\subsection{Method \code{normalizing_constant_power_parameter()}}{
Calculate the normalizing constant for the power parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$normalizing_constant_power_parameter(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Normalizing constant value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-inference"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-inference}{}}}
\subsection{Method \code{inference()}}{
Perform inference on the target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$inference(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string indicating the success status of the inference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-credible_interval"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-credible_interval}{}}}
\subsection{Method \code{credible_interval()}}{
Calculate the credible interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$credible_interval(level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{Credible interval level (default: 0.95).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector containing the lower and upper bounds of the credible interval.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-posterior_median"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-posterior_median}{}}}
\subsection{Method \code{posterior_median()}}{
Return the median of the posterior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$posterior_median(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Median of the posterior distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Sample from the posterior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$sample_posterior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of samples to draw from the posterior distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of samples from the posterior distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Sample from the prior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$sample_prior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of samples to draw from the prior distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of samples from the prior distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-posterior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-posterior_cdf}{}}}
\subsection{Method \code{posterior_cdf()}}{
Calculate the posterior cumulative distribution function (CDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$posterior_cdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Vector of points to evaluate the CDF at.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of CDF values corresponding to the input points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-posterior_pdf}{}}}
\subsection{Method \code{posterior_pdf()}}{
Calculate the posterior probability density function (PDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$posterior_pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Vector of points to evaluate the PDF at.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of PDF values corresponding to the input points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-prior_pdf}{}}}
\subsection{Method \code{prior_pdf()}}{
Calculate the prior probability density function (PDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$prior_pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Vector of points to evaluate the PDF at.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of PDF values corresponding to the input points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-plot_power_parameter_posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-plot_power_parameter_posterior_pdf}{}}}
\subsection{Method \code{plot_power_parameter_posterior_pdf()}}{
Plot posterior probability density function (PDF) of the power parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$plot_power_parameter_posterior_pdf(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-plot_power_parameter_vs_drift"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-plot_power_parameter_vs_drift}{}}}
\subsection{Method \code{plot_power_parameter_vs_drift()}}{
Plot the power parameter as a function of drift in treatment effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$plot_power_parameter_vs_drift(
  source_treatment_effect_estimate,
  target_data,
  min_drift,
  max_drift,
  resolution
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_treatment_effect_estimate}}{Treatment effect estimate in the source study}

\item{\code{target_data}}{Target study data}

\item{\code{min_drift}}{Minimum drift value}

\item{\code{max_drift}}{Maximum drift value}

\item{\code{resolution}}{Number of points on the drift grid.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prior PDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_NPP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_NPP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_NPP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
