% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_power_priors.R
\name{GaussianCommensuratePowerPrior}
\alias{GaussianCommensuratePowerPrior}
\title{GaussianCommensuratePowerPrior class}
\description{
This class represents a Gaussian Commensurate Power Prior model.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:MCMCModel]{RBExT::MCMCModel}} -> \code{GaussianCommensuratePowerPrior}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{Method name}

\item{\code{heterogeneity_prior_family}}{Heterogeneity prior family (half_normal, inverse_gamma)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianCommensuratePowerPrior-new}{\code{GaussianCommensuratePowerPrior$new()}}
\item \href{#method-GaussianCommensuratePowerPrior-prepare_data}{\code{GaussianCommensuratePowerPrior$prepare_data()}}
\item \href{#method-GaussianCommensuratePowerPrior-compute_posterior_parameters}{\code{GaussianCommensuratePowerPrior$compute_posterior_parameters()}}
\item \href{#method-GaussianCommensuratePowerPrior-sample_prior}{\code{GaussianCommensuratePowerPrior$sample_prior()}}
\item \href{#method-GaussianCommensuratePowerPrior-joint_prior_pdf}{\code{GaussianCommensuratePowerPrior$joint_prior_pdf()}}
\item \href{#method-GaussianCommensuratePowerPrior-unnormalized_prior_pdf}{\code{GaussianCommensuratePowerPrior$unnormalized_prior_pdf()}}
\item \href{#method-GaussianCommensuratePowerPrior-prior_pdf}{\code{GaussianCommensuratePowerPrior$prior_pdf()}}
\item \href{#method-GaussianCommensuratePowerPrior-clone}{\code{GaussianCommensuratePowerPrior$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="empirical_bayes_update"><a href='../../RBExT/html/Model.html#method-Model-empirical_bayes_update'><code>RBExT::Model$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_mean"><a href='../../RBExT/html/Model.html#method-Model-posterior_mean'><code>RBExT::Model$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_moments"><a href='../../RBExT/html/Model.html#method-Model-posterior_moments'><code>RBExT::Model$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-posterior_to_RBesT'><code>RBExT::Model$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_to_RBesT"><a href='../../RBExT/html/Model.html#method-Model-prior_to_RBesT'><code>RBExT::Model$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="check_data"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-check_data'><code>RBExT::MCMCModel$check_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="check_mcmc_config"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-check_mcmc_config'><code>RBExT::MCMCModel$check_mcmc_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="credible_interval"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-credible_interval'><code>RBExT::MCMCModel$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="draw_mcmc_prior"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-draw_mcmc_prior'><code>RBExT::MCMCModel$draw_mcmc_prior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="inference"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-inference'><code>RBExT::MCMCModel$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_cdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_cdf'><code>RBExT::MCMCModel$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_median"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_median'><code>RBExT::MCMCModel$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="posterior_pdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-posterior_pdf'><code>RBExT::MCMCModel$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="prior_cdf"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-prior_cdf'><code>RBExT::MCMCModel$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="MCMCModel" data-id="sample_posterior"><a href='../../RBExT/html/MCMCModel.html#method-MCMCModel-sample_posterior'><code>RBExT::MCMCModel$sample_posterior()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-new}{}}}
\subsection{Method \code{new()}}{
Initialize the GaussianCommensuratePowerPrior object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$new(prior, mcmc_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior object}

\item{\code{mcmc_config}}{The MCMC configuration parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
Prepare the data to be used by CmdStanR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$prepare_data(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-compute_posterior_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-compute_posterior_parameters}{}}}
\subsection{Method \code{compute_posterior_parameters()}}{
Compute posterior parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$compute_posterior_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Draw samples from the prior distribution. Based on equation 8 in Hobbs et al (2011).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$sample_prior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of samples}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-joint_prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-joint_prior_pdf}{}}}
\subsection{Method \code{joint_prior_pdf()}}{
Joint prior p.d.f. Based on equation (8) in Hobbs et al (2011).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$joint_prior_pdf(treatment_effect, gamma, tau)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatment_effect}}{Treatment effect}

\item{\code{gamma}}{Power parameter}

\item{\code{tau}}{Heterogeneity parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-unnormalized_prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-unnormalized_prior_pdf}{}}}
\subsection{Method \code{unnormalized_prior_pdf()}}{
Integrate out gamma and tau to get the marginal PDF for treatment_effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$unnormalized_prior_pdf(treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatment_effect}}{Treatment effect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-prior_pdf}{}}}
\subsection{Method \code{prior_pdf()}}{
Prior p.d.f. of the treatment effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$prior_pdf(treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatment_effect}}{Treatment effect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianCommensuratePowerPrior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianCommensuratePowerPrior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianCommensuratePowerPrior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
