% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_elastic_prior.R
\name{GaussianElasticPriorLogistic}
\alias{GaussianElasticPriorLogistic}
\title{GaussianElasticPriorLogistic class}
\description{
This class represents a Gaussian Elastic Prior model for logistic regression.
It inherits from the ConjugateGaussian class.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:ConjugateGaussian]{RBExT::ConjugateGaussian}} -> \code{GaussianElasticPriorLogistic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tuning_param_a}}{Numeric. Tuning parameter a for the elastic function.}

\item{\code{tuning_param_b}}{Numeric. Tuning parameter b for the elastic function.}

\item{\code{target_standard_error}}{Numeric. Standard error of the target study.}

\item{\code{target_treatment_effect_estimate}}{Numeric. Treatment effect estimate of the target study.}

\item{\code{target_sample_size_per_arm}}{Numeric. Sample size per arm in the target study.}

\item{\code{source_sample_size_per_arm}}{Numeric. Sample size per arm in the source study.}

\item{\code{source_data_variance}}{Numeric. Variance of the source data.}

\item{\code{method}}{Character. Method used for the analysis (default: "elastic_prior").}

\item{\code{shape_param}}{Numeric. Shape parameter for the elastic function.}

\item{\code{prior}}{List. Prior information for the analysis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GaussianElasticPriorLogistic-new}{\code{GaussianElasticPriorLogistic$new()}}
\item \href{#method-GaussianElasticPriorLogistic-empirical_bayes_update}{\code{GaussianElasticPriorLogistic$empirical_bayes_update()}}
\item \href{#method-GaussianElasticPriorLogistic-method_calibration}{\code{GaussianElasticPriorLogistic$method_calibration()}}
\item \href{#method-GaussianElasticPriorLogistic-determine_tuning_parameters}{\code{GaussianElasticPriorLogistic$determine_tuning_parameters()}}
\item \href{#method-GaussianElasticPriorLogistic-setup_prior}{\code{GaussianElasticPriorLogistic$setup_prior()}}
\item \href{#method-GaussianElasticPriorLogistic-clone}{\code{GaussianElasticPriorLogistic$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="inference"><a href='../../RBExT/html/Model.html#method-Model-inference'><code>RBExT::Model$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="credible_interval"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-credible_interval'><code>RBExT::ConjugateGaussian$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_cdf'><code>RBExT::ConjugateGaussian$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_mean"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_mean'><code>RBExT::ConjugateGaussian$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_median"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_median'><code>RBExT::ConjugateGaussian$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_moments"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_moments'><code>RBExT::ConjugateGaussian$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_pdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_pdf'><code>RBExT::ConjugateGaussian$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_to_RBesT'><code>RBExT::ConjugateGaussian$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_variance"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_variance'><code>RBExT::ConjugateGaussian$posterior_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_cdf'><code>RBExT::ConjugateGaussian$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_pdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_pdf'><code>RBExT::ConjugateGaussian$prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_to_RBesT'><code>RBExT::ConjugateGaussian$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_posterior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_posterior'><code>RBExT::ConjugateGaussian$sample_posterior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_prior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_prior'><code>RBExT::ConjugateGaussian$sample_prior()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-new"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-new}{}}}
\subsection{Method \code{new()}}{
Create an object from the GaussianElasticPriorLogistic class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$new(prior, summary_measure_likelihood)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{List. The prior information for the analysis.}

\item{\code{summary_measure_likelihood}}{List. Summary measure likelihood.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GaussianElasticPriorLogistic} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-empirical_bayes_update"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-empirical_bayes_update}{}}}
\subsection{Method \code{empirical_bayes_update()}}{
Updates the vague prior variance based on empirical Bayes approach.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$empirical_bayes_update(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{List. The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. Updates object fields internally.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-method_calibration"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-method_calibration}{}}}
\subsection{Method \code{method_calibration()}}{
Calibrates the method by determining tuning parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$method_calibration(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{List. The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. Updates tuning parameters internally.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-determine_tuning_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-determine_tuning_parameters}{}}}
\subsection{Method \code{determine_tuning_parameters()}}{
Determines tuning parameters a and b for the logistic elastic function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$determine_tuning_parameters(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{List. The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with components:
\itemize{
\item{a}{Numeric. Tuning parameter a in elastic function.}
\item{b}{Numeric. Tuning parameter b in elastic function.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-setup_prior"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-setup_prior}{}}}
\subsection{Method \code{setup_prior()}}{
Sets up the prior based on target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$setup_prior(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{List. The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None. Updates prior variance internally.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GaussianElasticPriorLogistic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GaussianElasticPriorLogistic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianElasticPriorLogistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
