% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_model.R
\name{Model}
\alias{Model}
\title{Model Class}
\description{
This class represents a statistical model for Bayesian borrowing analysis.
It provides methods for creating the model, performing inference, and calculating posterior moments.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{empirical_bayes}}{Logical indicating if empirical Bayes method is used.}

\item{\code{analytic_ocs}}{Results of the analytic operating characteristics simulation.}

\item{\code{posterior_parameters}}{Parameters of the posterior distribution.}

\item{\code{post_mean}}{Mean of the posterior distribution.}

\item{\code{post_median}}{Median of the posterior distribution.}

\item{\code{prior_mean}}{Mean of the prior distribution.}

\item{\code{prior_var}}{Variance of the prior distribution.}

\item{\code{post_var}}{Variance of the posterior distribution.}

\item{\code{parameters}}{List of model parameters.}

\item{\code{RBesT_prior}}{Prior distribution from RBesT package.}

\item{\code{RBesT_posterior}}{Posterior distribution from RBesT package.}

\item{\code{ess_morita}}{Effective sample size calculated using Morita's method.}

\item{\code{ess_elir}}{Effective sample size calculated using ELIR method.}

\item{\code{ess_moment}}{Effective sample size calculated using moment matching.}

\item{\code{prior}}{Prior distribution used in the model.}

\item{\code{method}}{Method used for the analysis.}

\item{\code{mcmc}}{Logical indicating if MCMC is used.}

\item{\code{RBesT_posterior_normix}}{Normal mixture approximation to the posterior distribution}

\item{\code{RBesT_prior_normix}}{Normal mixture approximation to the prior distribution}

\item{\code{summary_measure_likelihood}}{Likelihood familly, that is, the distribution used to model the summary measure of the treatment effect.}

\item{\code{n_components_mixture_approx}}{Number of mixture components used for the mixture approximation}

\item{\code{aic_penalty_parameter_mixture_approx}}{AIC penalty parameter used for the mixture approximation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model-new}{\code{Model$new()}}
\item \href{#method-Model-create}{\code{Model$create()}}
\item \href{#method-Model-inference}{\code{Model$inference()}}
\item \href{#method-Model-posterior_moments}{\code{Model$posterior_moments()}}
\item \href{#method-Model-prior_cdf}{\code{Model$prior_cdf()}}
\item \href{#method-Model-empirical_bayes_update}{\code{Model$empirical_bayes_update()}}
\item \href{#method-Model-posterior_mean}{\code{Model$posterior_mean()}}
\item \href{#method-Model-posterior_quantile}{\code{Model$posterior_quantile()}}
\item \href{#method-Model-posterior_median}{\code{Model$posterior_median()}}
\item \href{#method-Model-credible_interval}{\code{Model$credible_interval()}}
\item \href{#method-Model-prior_ESS}{\code{Model$prior_ESS()}}
\item \href{#method-Model-sample_prior}{\code{Model$sample_prior()}}
\item \href{#method-Model-sample_posterior}{\code{Model$sample_posterior()}}
\item \href{#method-Model-test_decision}{\code{Model$test_decision()}}
\item \href{#method-Model-simulation_for_given_treatment_effect}{\code{Model$simulation_for_given_treatment_effect()}}
\item \href{#method-Model-prior_treatment_benefit}{\code{Model$prior_treatment_benefit()}}
\item \href{#method-Model-estimate_frequentist_operating_characteristics}{\code{Model$estimate_frequentist_operating_characteristics()}}
\item \href{#method-Model-estimate_bayesian_operating_characteristics}{\code{Model$estimate_bayesian_operating_characteristics()}}
\item \href{#method-Model-prior_to_RBesT}{\code{Model$prior_to_RBesT()}}
\item \href{#method-Model-posterior_to_RBesT}{\code{Model$posterior_to_RBesT()}}
\item \href{#method-Model-plot_pdfs}{\code{Model$plot_pdfs()}}
\item \href{#method-Model-plot_prior_pdf}{\code{Model$plot_prior_pdf()}}
\item \href{#method-Model-plot_posterior_pdf}{\code{Model$plot_posterior_pdf()}}
\item \href{#method-Model-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Model object

Create a Model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-create"></a>}}
\if{latex}{\out{\hypertarget{method-Model-create}{}}}
\subsection{Method \code{create()}}{
This method creates a Model object based on the specified configuration and method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$create(
  case_study_config,
  method,
  method_parameters,
  source_data,
  mcmc_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case_study_config}}{A list containing the case study configuration.}

\item{\code{method}}{The method to be used for the analysis.}

\item{\code{method_parameters}}{A list containing the method parameters.}

\item{\code{source_data}}{Source data}

\item{\code{mcmc_config}}{An optional list containing the MCMC configuration.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Model object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-inference"></a>}}
\if{latex}{\out{\hypertarget{method-Model-inference}{}}}
\subsection{Method \code{inference()}}{
Perform inference on the Model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$inference(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data for the inference}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string "Success" if inference succeeded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-posterior_moments"></a>}}
\if{latex}{\out{\hypertarget{method-Model-posterior_moments}{}}}
\subsection{Method \code{posterior_moments()}}{
Abstract method to calculate the posterior moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$posterior_moments(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Data for which posterior moments need to be calculated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-prior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model-prior_cdf}{}}}
\subsection{Method \code{prior_cdf()}}{
Abstract method to calculate the cumulative distribution function of the prior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$prior_cdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The target treatment effect}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-empirical_bayes_update"></a>}}
\if{latex}{\out{\hypertarget{method-Model-empirical_bayes_update}{}}}
\subsection{Method \code{empirical_bayes_update()}}{
Method to update the prior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$empirical_bayes_update(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-posterior_mean"></a>}}
\if{latex}{\out{\hypertarget{method-Model-posterior_mean}{}}}
\subsection{Method \code{posterior_mean()}}{
Method to calculate the mean of the posterior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$posterior_mean(mcmc, n_samples_quantiles_estimation = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mcmc}}{Whether to use MCMC or not}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimates quantiles of distributions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The meen of the posterior distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-posterior_quantile"></a>}}
\if{latex}{\out{\hypertarget{method-Model-posterior_quantile}{}}}
\subsection{Method \code{posterior_quantile()}}{
Method to compute quantiles of the posterior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$posterior_quantile(p, mcmc, n_samples_quantiles_estimation = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Quantile}

\item{\code{mcmc}}{Whether to use MCMC estimation or not}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimates quantiles of distributions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The quantile of the posterior distribution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-posterior_median"></a>}}
\if{latex}{\out{\hypertarget{method-Model-posterior_median}{}}}
\subsection{Method \code{posterior_median()}}{
Method to calculate the median of the posterior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$posterior_median(mcmc, n_samples_quantiles_estimation = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mcmc}}{Whether to use MCMC or not}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimates quantiles of distributions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The median of the posterior distribution
Calculates the credible interval of the posterior distribution for a given level
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-credible_interval"></a>}}
\if{latex}{\out{\hypertarget{method-Model-credible_interval}{}}}
\subsection{Method \code{credible_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$credible_interval(
  level = 0.95,
  mcmc = FALSE,
  n_samples_quantiles_estimation = 10000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The level of confidence for the credible interval.}

\item{\code{mcmc}}{Whether to use MCMC or not}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimates quantiles of distributions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The credible interval.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-prior_ESS"></a>}}
\if{latex}{\out{\hypertarget{method-Model-prior_ESS}{}}}
\subsection{Method \code{prior_ESS()}}{
Method to calculate the effective sample size of the prior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$prior_ESS()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-Model-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Method to sample from the prior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$sample_prior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of samples to draw from the prior distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-Model-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Method to sample from the posterior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$sample_posterior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{Number of samples to draw from the posterior distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-test_decision"></a>}}
\if{latex}{\out{\hypertarget{method-Model-test_decision}{}}}
\subsection{Method \code{test_decision()}}{
This function returns the test decision for a fitted model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$test_decision(critical_value, theta_0, null_space, confidence_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{critical_value}}{The critical value for the test.}

\item{\code{theta_0}}{The null hypothesis value.}

\item{\code{null_space}}{The null space for the test.}

\item{\code{confidence_level}}{Confidence level for the test decision}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical value indicating the decision.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-simulation_for_given_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-Model-simulation_for_given_treatment_effect}{}}}
\subsection{Method \code{simulation_for_given_treatment_effect()}}{
Method to simulate for a given treatment effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$simulation_for_given_treatment_effect(
  target_data,
  n_replicates,
  critical_value,
  theta_0,
  confidence_level,
  null_space,
  case_study,
  method,
  to_return = c("credible_interval", "test_decision", "posterior_mean",
    "posterior_median", "ess_moment", "ess_precision", "mcmc_diagnostics"),
  verbose = 0,
  n_samples_quantiles_estimation
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Data for the target treatment effect}

\item{\code{n_replicates}}{Number of replicates to simulate}

\item{\code{critical_value}}{Critical value for hypothesis testing}

\item{\code{theta_0}}{Null hypothesis value}

\item{\code{confidence_level}}{Confidence level for the credible interval}

\item{\code{null_space}}{The null space for hypothesis testing}

\item{\code{case_study}}{Case study name}

\item{\code{method}}{Method name}

\item{\code{to_return}}{List of OCs to return}

\item{\code{verbose}}{Verbosity level (0 or 1)}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimate distributions quantiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of simulation results including test decisions, posterior means, medians, credible intervals, and posterior parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-prior_treatment_benefit"></a>}}
\if{latex}{\out{\hypertarget{method-Model-prior_treatment_benefit}{}}}
\subsection{Method \code{prior_treatment_benefit()}}{
Method to calculate the prior probability of treatment benefit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$prior_treatment_benefit(
  theta_0,
  null_space,
  n_prior_samples,
  tuning_length
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta_0}}{Null hypothesis value}

\item{\code{null_space}}{The null space for hypothesis testing}

\item{\code{n_prior_samples}}{Number of prior samples}

\item{\code{tuning_length}}{Tuning parameter length}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-estimate_frequentist_operating_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Model-estimate_frequentist_operating_characteristics}{}}}
\subsection{Method \code{estimate_frequentist_operating_characteristics()}}{
Method to estimate frequentist operating characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$estimate_frequentist_operating_characteristics(
  theta_0,
  target_data,
  n_replicates,
  critical_value,
  confidence_level,
  null_space,
  n_samples_quantiles_estimation,
  case_study,
  method,
  verbose = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta_0}}{Null hypothesis value}

\item{\code{target_data}}{Data for the target treatment effect}

\item{\code{n_replicates}}{Number of replicates to simulate}

\item{\code{critical_value}}{Critical value for hypothesis testing}

\item{\code{confidence_level}}{Confidence level for the credible interval}

\item{\code{null_space}}{The null space for hypothesis testing}

\item{\code{n_samples_quantiles_estimation}}{Number of samples used to estimate distributions quantiles.}

\item{\code{case_study}}{Case study name}

\item{\code{method}}{Method name}

\item{\code{verbose}}{Verbosity level (0 or 1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of estimated frequentist operating characteristics including coverage, MSE, bias, posterior mean, median, precision, credible intervals, and success probability
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-estimate_bayesian_operating_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Model-estimate_bayesian_operating_characteristics}{}}}
\subsection{Method \code{estimate_bayesian_operating_characteristics()}}{
Method to estimate Bayesian operating characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$estimate_bayesian_operating_characteristics(
  design_prior,
  theta_0,
  source_data,
  n_replicates,
  critical_value,
  confidence_level,
  null_space,
  n_samples_design_prior,
  target_sample_size_per_arm,
  case_study_config,
  target_to_source_std_ratio,
  simulation_config
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design_prior}}{Design prior}

\item{\code{theta_0}}{Null hypothesis value}

\item{\code{source_data}}{Source data}

\item{\code{n_replicates}}{Number of replicates to simulate}

\item{\code{critical_value}}{Critical value for hypothesis testing}

\item{\code{confidence_level}}{Confidence level for the credible interval}

\item{\code{null_space}}{The null space for hypothesis testing}

\item{\code{n_samples_design_prior}}{Number of samples from the design prior}

\item{\code{target_sample_size_per_arm}}{Sample size per arm in the target study}

\item{\code{case_study_config}}{Configuration of the case study}

\item{\code{target_to_source_std_ratio}}{Ratio between the target and source study standard deviation}

\item{\code{simulation_config}}{Simulation configuration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of estimated frequentist operating characteristics including coverage, MSE, bias, posterior mean, median, precision, credible intervals, and success probability
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-prior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-Model-prior_to_RBesT}{}}}
\subsection{Method \code{prior_to_RBesT()}}{
Method to convert the model to RBesT
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$prior_to_RBesT(n_samples_mixture_approx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples_mixture_approx}}{Number of samples used to approximate the prior with a mixture.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-posterior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-Model-posterior_to_RBesT}{}}}
\subsection{Method \code{posterior_to_RBesT()}}{
Convert the posterior distribution to RBesT format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$posterior_to_RBesT(target_data, simulation_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}

\item{\code{simulation_config}}{Configuration of simulation study}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-plot_pdfs"></a>}}
\if{latex}{\out{\hypertarget{method-Model-plot_pdfs}{}}}
\subsection{Method \code{plot_pdfs()}}{
Plot prior and posterior probability density functions (PDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$plot_pdfs(xmin = -2, xmax = 2, resolution = 100, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmin}}{Lower bound of the treatment effect range}

\item{\code{xmax}}{Upper bound of the treatment effect range}

\item{\code{resolution}}{Resolution of the treatment effect range}

\item{\code{...}}{Optional argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-plot_prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model-plot_prior_pdf}{}}}
\subsection{Method \code{plot_prior_pdf()}}{
Plot prior probability density function (PDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$plot_prior_pdf(xmin = -2, xmax = 2, resolution = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmin}}{Lower bound of the treatment effect range}

\item{\code{xmax}}{Upper bound of the treatment effect range}

\item{\code{resolution}}{Resolution of the treatment effect range}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-plot_posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model-plot_posterior_pdf}{}}}
\subsection{Method \code{plot_posterior_pdf()}}{
Plot posterior probability density function (PDF).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$plot_posterior_pdf(xmin = -2, xmax = 2, resolution = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmin}}{Lower bound of the treatment effect range}

\item{\code{xmax}}{Upper bound of the treatment effect range}

\item{\code{resolution}}{Resolution of the treatment effect range}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
