% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_model.R
\name{Model_RBesT}
\alias{Model_RBesT}
\title{Model_RBesT}
\description{
An R6 class representing a Bayesian model using RBesT.
}
\section{Super class}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{Model_RBesT}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{posterior_summary}}{Summary of the posterior distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model_RBesT-new}{\code{Model_RBesT$new()}}
\item \href{#method-Model_RBesT-prior_pdf}{\code{Model_RBesT$prior_pdf()}}
\item \href{#method-Model_RBesT-prior_cdf}{\code{Model_RBesT$prior_cdf()}}
\item \href{#method-Model_RBesT-posterior_moments}{\code{Model_RBesT$posterior_moments()}}
\item \href{#method-Model_RBesT-posterior_mean}{\code{Model_RBesT$posterior_mean()}}
\item \href{#method-Model_RBesT-posterior_variance}{\code{Model_RBesT$posterior_variance()}}
\item \href{#method-Model_RBesT-posterior_median}{\code{Model_RBesT$posterior_median()}}
\item \href{#method-Model_RBesT-posterior_pdf}{\code{Model_RBesT$posterior_pdf()}}
\item \href{#method-Model_RBesT-posterior_cdf}{\code{Model_RBesT$posterior_cdf()}}
\item \href{#method-Model_RBesT-sample_prior}{\code{Model_RBesT$sample_prior()}}
\item \href{#method-Model_RBesT-sample_posterior}{\code{Model_RBesT$sample_posterior()}}
\item \href{#method-Model_RBesT-prior_to_RBesT}{\code{Model_RBesT$prior_to_RBesT()}}
\item \href{#method-Model_RBesT-posterior_to_RBesT}{\code{Model_RBesT$posterior_to_RBesT()}}
\item \href{#method-Model_RBesT-credible_interval}{\code{Model_RBesT$credible_interval()}}
\item \href{#method-Model_RBesT-clone}{\code{Model_RBesT$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="empirical_bayes_update"><a href='../../RBExT/html/Model.html#method-Model-empirical_bayes_update'><code>RBExT::Model$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="inference"><a href='../../RBExT/html/Model.html#method-Model-inference'><code>RBExT::Model$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-new}{}}}
\subsection{Method \code{new()}}{
Initializes the GaussianRMP object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$new(prior)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior information for the analysis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-prior_pdf}{}}}
\subsection{Method \code{prior_pdf()}}{
Calculates the prior probability density function (PDF) for a given target treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$prior_pdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The target treatment effect.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prior PDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-prior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-prior_cdf}{}}}
\subsection{Method \code{prior_cdf()}}{
Calculates the prior cumulative distribution function (CDF) for a given target treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$prior_cdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The target treatment effect.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prior CDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_moments"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_moments}{}}}
\subsection{Method \code{posterior_moments()}}{
Calculates the posterior moments based on the target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_moments(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data for the analysis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_mean"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_mean}{}}}
\subsection{Method \code{posterior_mean()}}{
Calculates the posterior mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The posterior mean.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_variance"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_variance}{}}}
\subsection{Method \code{posterior_variance()}}{
Calculates the posterior variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_variance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The posterior variance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_median"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_median}{}}}
\subsection{Method \code{posterior_median()}}{
Calculates the posterior median.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_median(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The posterior median.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_pdf}{}}}
\subsection{Method \code{posterior_pdf()}}{
Calculates the posterior probability density function (PDF) for a given target treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_pdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The target treatment effect.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The posterior PDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_cdf}{}}}
\subsection{Method \code{posterior_cdf()}}{
Calculates the posterior cumulative distribution function (CDF) for a given target treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_cdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The target treatment effect.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The posterior CDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Samples from the prior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$sample_prior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The samples from the prior distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Samples from the posterior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$sample_posterior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The samples from the posterior distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-prior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-prior_to_RBesT}{}}}
\subsection{Method \code{prior_to_RBesT()}}{
Converts the prior distribution to the RBesT format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$prior_to_RBesT(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-posterior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-posterior_to_RBesT}{}}}
\subsection{Method \code{posterior_to_RBesT()}}{
Converts the posterior distribution to the RBesT format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$posterior_to_RBesT(target_data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-credible_interval"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-credible_interval}{}}}
\subsection{Method \code{credible_interval()}}{
Calculates the credible interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$credible_interval(level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{Level of the credible interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector containing the lower and upper bounds of the credible interval.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model_RBesT-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model_RBesT-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model_RBesT$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
