% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{ObservedSourceData}
\alias{ObservedSourceData}
\title{ObservedSourceData class}
\description{
This class represents the observed source data used for Bayesian borrowing in clinical studies.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endpoint}}{character. The endpoint of the study. Possible values are "continuous", "binary", "time_to_event", "recurrent_event".}

\item{\code{summary_measure_likelihood}}{character. The distribution of the summary measure. Possible values are "normal", "Binomial".}

\item{\code{sample_size_control}}{numeric. The sample size in the control arm of the source data.}

\item{\code{sample_size_treatment}}{numeric. The sample size in the treatment arm of the source data.}

\item{\code{treatment_effect_estimate}}{numeric. The estimated treatment effect in the source data.}

\item{\code{standard_error}}{numeric. The standard error of the treatment effect estimate in the source data.}

\item{\code{control_rate}}{numeric. The rate of success (1s) in the control arm of the source data (for binary endpoint).}

\item{\code{treatment_rate}}{numeric. The rate of success (1s) in the treatment arm of the source data (for binary endpoint).}

\item{\code{equivalent_source_sample_size_per_arm}}{numeric. Equivalent sample size per arm.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservedSourceData-new}{\code{ObservedSourceData$new()}}
\item \href{#method-ObservedSourceData-to_dict}{\code{ObservedSourceData$to_dict()}}
\item \href{#method-ObservedSourceData-clone}{\code{ObservedSourceData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedSourceData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedSourceData-new}{}}}
\subsection{Method \code{new()}}{
Initialize an instance of ObservedSourceData
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedSourceData$new(case_study_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case_study_config}}{list. A list containing the configuration for the case study.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ObservedSourceData.
Convert the object's data to a dictionary format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedSourceData-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedSourceData-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedSourceData$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list. A list containing the source data attributes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservedSourceData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservedSourceData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservedSourceData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
