% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_empirical_bayes_power_prior.R
\name{p_value_based_PP_Gaussian}
\alias{p_value_based_PP_Gaussian}
\title{p_value_based_PP_Gaussian class}
\description{
This class represents a p-value based power prior method.
It inherits from the Gaussian_empirical_Bayes_PP class.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:ConjugateGaussian]{RBExT::ConjugateGaussian}} -> \code{\link[RBExT:StaticBorrowingGaussian]{RBExT::StaticBorrowingGaussian}} -> \code{\link[RBExT:Gaussian_empirical_Bayes_PP]{RBExT::Gaussian_empirical_Bayes_PP}} -> \code{p_value_based_PP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{shape_parameter}}{The shape parameter for the method.}

\item{\code{equivalence_margin}}{The equivalence margin for the method.}

\item{\code{method}}{Method name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-p_value_based_PP-new}{\code{p_value_based_PP_Gaussian$new()}}
\item \href{#method-p_value_based_PP-test}{\code{p_value_based_PP_Gaussian$test()}}
\item \href{#method-p_value_based_PP-power_parameter_estimation}{\code{p_value_based_PP_Gaussian$power_parameter_estimation()}}
\item \href{#method-p_value_based_PP-clone}{\code{p_value_based_PP_Gaussian$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="test_decision"><a href='../../RBExT/html/Model.html#method-Model-test_decision'><code>RBExT::Model$test_decision()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="credible_interval"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-credible_interval'><code>RBExT::ConjugateGaussian$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_cdf'><code>RBExT::ConjugateGaussian$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_mean"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_mean'><code>RBExT::ConjugateGaussian$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_median"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_median'><code>RBExT::ConjugateGaussian$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_moments"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_moments'><code>RBExT::ConjugateGaussian$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_pdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_pdf'><code>RBExT::ConjugateGaussian$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_to_RBesT'><code>RBExT::ConjugateGaussian$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="posterior_variance"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-posterior_variance'><code>RBExT::ConjugateGaussian$posterior_variance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_cdf"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_cdf'><code>RBExT::ConjugateGaussian$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="prior_to_RBesT"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-prior_to_RBesT'><code>RBExT::ConjugateGaussian$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_posterior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_posterior'><code>RBExT::ConjugateGaussian$sample_posterior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="ConjugateGaussian" data-id="sample_prior"><a href='../../RBExT/html/ConjugateGaussian.html#method-ConjugateGaussian-sample_prior'><code>RBExT::ConjugateGaussian$sample_prior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Gaussian_empirical_Bayes_PP" data-id="empirical_bayes_update"><a href='../../RBExT/html/Gaussian_empirical_Bayes_PP.html#method-Gaussian_empirical_Bayes_PP-empirical_bayes_update'><code>RBExT::Gaussian_empirical_Bayes_PP$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Gaussian_empirical_Bayes_PP" data-id="hypothesis_space_transformation"><a href='../../RBExT/html/Gaussian_empirical_Bayes_PP.html#method-Gaussian_empirical_Bayes_PP-hypothesis_space_transformation'><code>RBExT::Gaussian_empirical_Bayes_PP$hypothesis_space_transformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Gaussian_empirical_Bayes_PP" data-id="inference"><a href='../../RBExT/html/Gaussian_empirical_Bayes_PP.html#method-Gaussian_empirical_Bayes_PP-inference'><code>RBExT::Gaussian_empirical_Bayes_PP$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Gaussian_empirical_Bayes_PP" data-id="plot_power_parameter_vs_drift"><a href='../../RBExT/html/Gaussian_empirical_Bayes_PP.html#method-Gaussian_empirical_Bayes_PP-plot_power_parameter_vs_drift'><code>RBExT::Gaussian_empirical_Bayes_PP$plot_power_parameter_vs_drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Gaussian_empirical_Bayes_PP" data-id="prior_pdf"><a href='../../RBExT/html/Gaussian_empirical_Bayes_PP.html#method-Gaussian_empirical_Bayes_PP-prior_pdf'><code>RBExT::Gaussian_empirical_Bayes_PP$prior_pdf()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP-new"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP-new}{}}}
\subsection{Method \code{new()}}{
Initialize the p_value_based_PP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{p_value_based_PP_Gaussian$new(prior, theta_0, null_space)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior object.}

\item{\code{theta_0}}{Boundary of the null hypothesis space}

\item{\code{null_space}}{Null space.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
Test method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP-test"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP-test}{}}}
\subsection{Method \code{test()}}{
This method performs the test for the given target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{p_value_based_PP_Gaussian$test(
  target_data,
  source_treatment_effect_estimate,
  target_treatment_effect_estimate,
  test_type = "t-test"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data object.}

\item{\code{source_treatment_effect_estimate}}{The source treatment effect estimate.}

\item{\code{target_treatment_effect_estimate}}{The target treatment effect estimate.}

\item{\code{test_type}}{Type of frequentist test used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The p-value.
Power parameter estimation method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP-power_parameter_estimation"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP-power_parameter_estimation}{}}}
\subsection{Method \code{power_parameter_estimation()}}{
This method estimates the power parameter for the given target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{p_value_based_PP_Gaussian$power_parameter_estimation(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The target data object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The power parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p_value_based_PP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-p_value_based_PP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{p_value_based_PP_Gaussian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
