% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{simulation_scenarios}
\alias{simulation_scenarios}
\title{Generate simulation scenarios based on the configuration directory.}
\usage{
simulation_scenarios(config_dir, scenarios_config)
}
\arguments{
\item{config_dir}{The directory containing the configuration files.}

\item{scenarios_config}{The simulation configuration file.}
}
\value{
A list of simulation scenarios.
}
\description{
Generate simulation scenarios based on the configuration directory.
}
