% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_compute_sweet_spot.R
\name{sweet_spot}
\alias{sweet_spot}
\title{Calculate Sweet Spots for Multiple Metrics}
\usage{
sweet_spot(results_freq_df, metrics, based_on_CI = TRUE)
}
\arguments{
\item{results_freq_df}{A dataframe containing the results frequency data with columns
"case_study", "method", "source_denominator_change_factor", "control_drift", and "target_sample_size_per_arm".}

\item{metrics}{A list of metrics to evaluate. Each metric should have a \code{name} attribute.}
}
\value{
A dataframe containing the sweet spots for each metric, case study, method, and combination
of parameters.
}
\description{
This function calculates the sweet spots for various metrics over multiple case studies and methods.
It determines where the metric differences between two methods cross a reference value.
}
\examples{
\dontrun{
results <- sweet_spot(results_freq_df, metrics)
}
}
