% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{TargetData}
\alias{TargetData}
\title{TargetData class}
\description{
A base class for target data objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endpoint}}{The endpoint of the target data.}

\item{\code{treatment_drift}}{The treatment drift value.}

\item{\code{control_drift}}{The control drift value.}

\item{\code{drift}}{The drift value calculated as the difference between treatment and control drifts.}

\item{\code{summary_measure_likelihood}}{The distribution of the summary measure.}

\item{\code{sample_size_per_arm}}{The target sample size per arm.}

\item{\code{sample_size_control}}{The target sample size for the control group.}

\item{\code{sample_size_treatment}}{The target sample size for the treatment group.}

\item{\code{treatment_effect}}{The target treatment effect value.}

\item{\code{sampling_approximation}}{The flag indicating if sampling approximation is used.}

\item{\code{standard_deviation}}{The target standard deviation value.}

\item{\code{sample}}{The sample data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TargetData-new}{\code{TargetData$new()}}
\item \href{#method-TargetData-to_dict}{\code{TargetData$to_dict()}}
\item \href{#method-TargetData-plot_sample}{\code{TargetData$plot_sample()}}
\item \href{#method-TargetData-clone}{\code{TargetData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TargetData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TargetData-new}{}}}
\subsection{Method \code{new()}}{
Initializes the target data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TargetData$new(
  source_data,
  sampling_approximation,
  target_sample_size_per_arm,
  control_drift = 0,
  treatment_drift,
  summary_measure_likelihood
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_data}}{The source data object.}

\item{\code{sampling_approximation}}{The sampling approximation flag.}

\item{\code{target_sample_size_per_arm}}{The target sample size per arm.}

\item{\code{control_drift}}{The control drift.}

\item{\code{treatment_drift}}{The treatment drift.}

\item{\code{summary_measure_likelihood}}{The summary measure distribution.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TargetData-to_dict"></a>}}
\if{latex}{\out{\hypertarget{method-TargetData-to_dict}{}}}
\subsection{Method \code{to_dict()}}{
Converts the target data object to a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TargetData$to_dict()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing the target data object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TargetData-plot_sample"></a>}}
\if{latex}{\out{\hypertarget{method-TargetData-plot_sample}{}}}
\subsection{Method \code{plot_sample()}}{
Plot the target data samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TargetData$plot_sample(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data samples}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TargetData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TargetData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TargetData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
