% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_test_then_pool.R
\name{TestThenPool}
\alias{TestThenPool}
\title{TestThenPool class}
\description{
This class implements the Test-Then-Pool framework for Bayesian borrowing in clinical trials.
}
\details{
The TestThenPool class provides methods for testing and inference with the Test-then-Pool method.
}
\section{Super class}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{TestThenPool}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pooling}}{Pooling model}

\item{\code{separate}}{Separate model}

\item{\code{pool}}{Pooling indicator}

\item{\code{inference_method}}{Inference method}

\item{\code{summary_measure_likelihood}}{Summary measure distribution}

\item{\code{source_treatment_effect_estimate}}{Source treatment effect estimate}

\item{\code{source_standard_error}}{Source standard error}

\item{\code{method}}{Method name}

\item{\code{empirical_bayes}}{Indicator that the method uses empirical Bayes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestThenPool-new}{\code{TestThenPool$new()}}
\item \href{#method-TestThenPool-test}{\code{TestThenPool$test()}}
\item \href{#method-TestThenPool-test_pvalue}{\code{TestThenPool$test_pvalue()}}
\item \href{#method-TestThenPool-inference}{\code{TestThenPool$inference()}}
\item \href{#method-TestThenPool-empirical_bayes_update}{\code{TestThenPool$empirical_bayes_update()}}
\item \href{#method-TestThenPool-credible_interval}{\code{TestThenPool$credible_interval()}}
\item \href{#method-TestThenPool-posterior_median}{\code{TestThenPool$posterior_median()}}
\item \href{#method-TestThenPool-posterior_pdf}{\code{TestThenPool$posterior_pdf()}}
\item \href{#method-TestThenPool-prior_pdf}{\code{TestThenPool$prior_pdf()}}
\item \href{#method-TestThenPool-posterior_cdf}{\code{TestThenPool$posterior_cdf()}}
\item \href{#method-TestThenPool-prior_cdf}{\code{TestThenPool$prior_cdf()}}
\item \href{#method-TestThenPool-sample_prior}{\code{TestThenPool$sample_prior()}}
\item \href{#method-TestThenPool-sample_posterior}{\code{TestThenPool$sample_posterior()}}
\item \href{#method-TestThenPool-prior_to_RBesT}{\code{TestThenPool$prior_to_RBesT()}}
\item \href{#method-TestThenPool-posterior_to_RBesT}{\code{TestThenPool$posterior_to_RBesT()}}
\item \href{#method-TestThenPool-test_decision}{\code{TestThenPool$test_decision()}}
\item \href{#method-TestThenPool-plot_test_vs_drift}{\code{TestThenPool$plot_test_vs_drift()}}
\item \href{#method-TestThenPool-plot_test_pvalue_vs_drift}{\code{TestThenPool$plot_test_pvalue_vs_drift()}}
\item \href{#method-TestThenPool-clone}{\code{TestThenPool$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_mean"><a href='../../RBExT/html/Model.html#method-Model-posterior_mean'><code>RBExT::Model$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_moments"><a href='../../RBExT/html/Model.html#method-Model-posterior_moments'><code>RBExT::Model$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-new"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-new}{}}}
\subsection{Method \code{new()}}{
Initializes a TestThenPool object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$new(prior, mcmc_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{The prior distribution for the treatment effect.}

\item{\code{mcmc_config}}{Configuration for the MCMC sampling.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-test"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-test}{}}}
\subsection{Method \code{test()}}{
Performs the test with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$test(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The data from the target study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-test_pvalue"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-test_pvalue}{}}}
\subsection{Method \code{test_pvalue()}}{
p-value of the test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$test_pvalue(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The data from the target study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-inference"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-inference}{}}}
\subsection{Method \code{inference()}}{
Performs inference with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$inference(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The data from the target study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-empirical_bayes_update"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-empirical_bayes_update}{}}}
\subsection{Method \code{empirical_bayes_update()}}{
Performs the empirical Bayes update with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$empirical_bayes_update(target_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{The data from the target study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-credible_interval"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-credible_interval}{}}}
\subsection{Method \code{credible_interval()}}{
Calculates the credible interval with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$credible_interval(level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{The confidence level for the credible interval (default is 0.95).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-posterior_median"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-posterior_median}{}}}
\subsection{Method \code{posterior_median()}}{
Return the median of the posterior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$posterior_median(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-posterior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-posterior_pdf}{}}}
\subsection{Method \code{posterior_pdf()}}{
Calculates the posterior probability density function with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$posterior_pdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The treatment effect of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-prior_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-prior_pdf}{}}}
\subsection{Method \code{prior_pdf()}}{
Calculates the prior probability density function with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$prior_pdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The treatment effect of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-posterior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-posterior_cdf}{}}}
\subsection{Method \code{posterior_cdf()}}{
Calculates the posterior cumulative distribution function with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$posterior_cdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{The treatment effect of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-prior_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-prior_cdf}{}}}
\subsection{Method \code{prior_cdf()}}{
CDF of the prior distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$prior_cdf(target_treatment_effect)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_treatment_effect}}{Treatment effect value in the target study.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-sample_prior"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-sample_prior}{}}}
\subsection{Method \code{sample_prior()}}{
Samples from the prior distribution with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$sample_prior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-sample_posterior"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-sample_posterior}{}}}
\subsection{Method \code{sample_posterior()}}{
Samples from the posterior distribution with the Test-then-Pool method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$sample_posterior(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-prior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-prior_to_RBesT}{}}}
\subsection{Method \code{prior_to_RBesT()}}{
Converts the prior distribution to the RBesT format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$prior_to_RBesT(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-posterior_to_RBesT"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-posterior_to_RBesT}{}}}
\subsection{Method \code{posterior_to_RBesT()}}{
Convert the posterior distribution to RBesT format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$posterior_to_RBesT(target_data, simulation_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}

\item{\code{simulation_config}}{Simulation configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-test_decision"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-test_decision}{}}}
\subsection{Method \code{test_decision()}}{
Return the test decision based on the posterior distribution. The decision rule is: \eqn{P(\theta_T > \theta_0 \mid \mathbf{D}_S, \mathbf{D}_T) > ) \eta} (if null_space is right).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$test_decision(
  critical_value,
  theta_0,
  null_space,
  confidence_level
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{critical_value}}{Critical value}

\item{\code{theta_0}}{Boundary of the null hypothesis space}

\item{\code{null_space}}{Side of the null hypothesis space}

\item{\code{confidence_level}}{Confidence level of the test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-plot_test_vs_drift"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-plot_test_vs_drift}{}}}
\subsection{Method \code{plot_test_vs_drift()}}{
Plot the pooling test decision as a function of drift in treatment effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$plot_test_vs_drift(
  source_treatment_effect_estimate,
  target_data,
  min_drift,
  max_drift,
  resolution
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_treatment_effect_estimate}}{Treatment effect estimate in the source study}

\item{\code{target_data}}{Target study data}

\item{\code{min_drift}}{Minimum drift value}

\item{\code{max_drift}}{Maximum drift value}

\item{\code{resolution}}{Number of points on the drift grid.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prior PDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-plot_test_pvalue_vs_drift"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-plot_test_pvalue_vs_drift}{}}}
\subsection{Method \code{plot_test_pvalue_vs_drift()}}{
Plot the pooling test p-value as a function of drift in treatment effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$plot_test_pvalue_vs_drift(
  source_treatment_effect_estimate,
  target_data,
  min_drift,
  max_drift,
  resolution
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_treatment_effect_estimate}}{Treatment effect estimate in the source study}

\item{\code{target_data}}{Target study data}

\item{\code{min_drift}}{Minimum drift value}

\item{\code{max_drift}}{Maximum drift value}

\item{\code{resolution}}{Number of points on the drift grid.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prior PDF.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPool-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPool-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
