% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_test_then_pool.R
\name{TestThenPoolDifference}
\alias{TestThenPoolDifference}
\title{TestThenPoolDifference class}
\description{
This class extends the TestThenPool class to implement the Test-Then-Pool framework for difference trials.
}
\details{
The TestThenPoolDifference class provides methods for testing and inference in difference trials using the Test-Then-Pool framework.
}
\section{Super classes}{
\code{\link[RBExT:Model]{RBExT::Model}} -> \code{\link[RBExT:TestThenPool]{RBExT::TestThenPool}} -> \code{TestThenPoolDifference}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{significance_level}}{Significance level for the test}

\item{\code{method}}{Method name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestThenPoolDifference-new}{\code{TestThenPoolDifference$new()}}
\item \href{#method-TestThenPoolDifference-test_pvalue}{\code{TestThenPoolDifference$test_pvalue()}}
\item \href{#method-TestThenPoolDifference-test}{\code{TestThenPoolDifference$test()}}
\item \href{#method-TestThenPoolDifference-clone}{\code{TestThenPoolDifference$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="create"><a href='../../RBExT/html/Model.html#method-Model-create'><code>RBExT::Model$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_bayesian_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_bayesian_operating_characteristics'><code>RBExT::Model$estimate_bayesian_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="estimate_frequentist_operating_characteristics"><a href='../../RBExT/html/Model.html#method-Model-estimate_frequentist_operating_characteristics'><code>RBExT::Model$estimate_frequentist_operating_characteristics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_pdfs"><a href='../../RBExT/html/Model.html#method-Model-plot_pdfs'><code>RBExT::Model$plot_pdfs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_posterior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_posterior_pdf'><code>RBExT::Model$plot_posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="plot_prior_pdf"><a href='../../RBExT/html/Model.html#method-Model-plot_prior_pdf'><code>RBExT::Model$plot_prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_mean"><a href='../../RBExT/html/Model.html#method-Model-posterior_mean'><code>RBExT::Model$posterior_mean()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_moments"><a href='../../RBExT/html/Model.html#method-Model-posterior_moments'><code>RBExT::Model$posterior_moments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="posterior_quantile"><a href='../../RBExT/html/Model.html#method-Model-posterior_quantile'><code>RBExT::Model$posterior_quantile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_ESS"><a href='../../RBExT/html/Model.html#method-Model-prior_ESS'><code>RBExT::Model$prior_ESS()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="prior_treatment_benefit"><a href='../../RBExT/html/Model.html#method-Model-prior_treatment_benefit'><code>RBExT::Model$prior_treatment_benefit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="Model" data-id="simulation_for_given_treatment_effect"><a href='../../RBExT/html/Model.html#method-Model-simulation_for_given_treatment_effect'><code>RBExT::Model$simulation_for_given_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="credible_interval"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-credible_interval'><code>RBExT::TestThenPool$credible_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="empirical_bayes_update"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-empirical_bayes_update'><code>RBExT::TestThenPool$empirical_bayes_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="inference"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-inference'><code>RBExT::TestThenPool$inference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="plot_test_pvalue_vs_drift"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-plot_test_pvalue_vs_drift'><code>RBExT::TestThenPool$plot_test_pvalue_vs_drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="plot_test_vs_drift"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-plot_test_vs_drift'><code>RBExT::TestThenPool$plot_test_vs_drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="posterior_cdf"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-posterior_cdf'><code>RBExT::TestThenPool$posterior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="posterior_median"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-posterior_median'><code>RBExT::TestThenPool$posterior_median()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="posterior_pdf"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-posterior_pdf'><code>RBExT::TestThenPool$posterior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="posterior_to_RBesT"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-posterior_to_RBesT'><code>RBExT::TestThenPool$posterior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="prior_cdf"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-prior_cdf'><code>RBExT::TestThenPool$prior_cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="prior_pdf"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-prior_pdf'><code>RBExT::TestThenPool$prior_pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="prior_to_RBesT"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-prior_to_RBesT'><code>RBExT::TestThenPool$prior_to_RBesT()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="sample_posterior"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-sample_posterior'><code>RBExT::TestThenPool$sample_posterior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="sample_prior"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-sample_prior'><code>RBExT::TestThenPool$sample_prior()</code></a></span></li>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="TestThenPool" data-id="test_decision"><a href='../../RBExT/html/TestThenPool.html#method-TestThenPool-test_decision'><code>RBExT::TestThenPool$test_decision()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPoolDifference-new"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPoolDifference-new}{}}}
\subsection{Method \code{new()}}{
Instantiate model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPoolDifference$new(prior, mcmc_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior}}{Prior}

\item{\code{mcmc_config}}{MCMC configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPoolDifference-test_pvalue"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPoolDifference-test_pvalue}{}}}
\subsection{Method \code{test_pvalue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPoolDifference$test_pvalue(target_data, test_type = "t-test")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}

\item{\code{test_type}}{Frequentist test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPoolDifference-test"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPoolDifference-test}{}}}
\subsection{Method \code{test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPoolDifference$test(target_data, test_type = "t-test")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_data}}{Target study data}

\item{\code{test_type}}{Frequentist test}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestThenPoolDifference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestThenPoolDifference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestThenPoolDifference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
