% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_design_prior.R
\name{UnitInformationDesignPrior}
\alias{UnitInformationDesignPrior}
\title{UnitInformationDesignPrior class}
\description{
A class representing the unit information design prior for Bayesian borrowing.
}
\section{Super class}{
\code{\link[RBExT:DesignPrior]{RBExT::DesignPrior}} -> \code{UnitInformationDesignPrior}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameters}}{Parameters for the prior}

\item{\code{summary_measure_likelihood}}{Treatment effect distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnitInformationDesignPrior-new}{\code{UnitInformationDesignPrior$new()}}
\item \href{#method-UnitInformationDesignPrior-sample}{\code{UnitInformationDesignPrior$sample()}}
\item \href{#method-UnitInformationDesignPrior-cdf}{\code{UnitInformationDesignPrior$cdf()}}
\item \href{#method-UnitInformationDesignPrior-pdf}{\code{UnitInformationDesignPrior$pdf()}}
\item \href{#method-UnitInformationDesignPrior-clone}{\code{UnitInformationDesignPrior$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="RBExT" data-topic="DesignPrior" data-id="create"><a href='../../RBExT/html/DesignPrior.html#method-DesignPrior-create'><code>RBExT::DesignPrior$create()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitInformationDesignPrior-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnitInformationDesignPrior-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new instance of UnitInformationDesignPrior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitInformationDesignPrior$new(
  source_data,
  case_study_config,
  case_study,
  simulation_config,
  mcmc_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_data}}{The source data object.}

\item{\code{case_study_config}}{Case study configuration.}

\item{\code{case_study}}{Case study name}

\item{\code{simulation_config}}{Simulation config}

\item{\code{mcmc_config}}{MCMC configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitInformationDesignPrior-sample"></a>}}
\if{latex}{\out{\hypertarget{method-UnitInformationDesignPrior-sample}{}}}
\subsection{Method \code{sample()}}{
Samples from the unit information design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitInformationDesignPrior$sample(n_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_samples}}{The number of samples to generate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitInformationDesignPrior-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-UnitInformationDesignPrior-cdf}{}}}
\subsection{Method \code{cdf()}}{
Computes the cumulative distribution function (CDF) of the unit information design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitInformationDesignPrior$cdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the CDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitInformationDesignPrior-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-UnitInformationDesignPrior-pdf}{}}}
\subsection{Method \code{pdf()}}{
Computes the cumulative distribution function (CDF) of the unit information design prior.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitInformationDesignPrior$pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value at which to evaluate the CDF.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitInformationDesignPrior-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnitInformationDesignPrior-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitInformationDesignPrior$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
