% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_bayesian_ocs.R
\name{upper_bound_proba_FP_MC}
\alias{upper_bound_proba_FP_MC}
\title{Calculate the upper bound probability of false positive}
\usage{
upper_bound_proba_FP_MC(
  model,
  prior_proba_no_benefit,
  source_data,
  theta_0,
  target_sample_size_per_arm,
  case_study_config,
  n_replicates,
  confidence_level,
  null_space,
  critical_value
)
}
\arguments{
\item{model}{The model.}

\item{prior_proba_no_benefit}{The prior probability of no benefit.}

\item{source_data}{The source data.}

\item{theta_0}{The value of theta_0.}

\item{target_sample_size_per_arm}{The target sample size per arm.}

\item{case_study_config}{The case study configuration.}

\item{n_replicates}{The number of replicates.}

\item{confidence_level}{The confidence level.}

\item{null_space}{The null space (either "left" or "right").}

\item{critical_value}{The critical value.}
}
\value{
The upper bound probability of false positive, defined as \eqn{Pr(Study success|\theta_T = \theta_0) \times Pr(\theta_T \leq \theta_0)}
}
\description{
This function calculates the upper bound probability of false positive based on the model, prior probability of no benefit,
source data, theta_0, target sample size per arm, case study configuration, number of replicates, confidence level,
null space, and critical value.
}
\examples{
NA

}
